@@ -5,58 +5,60 @@ publish_to: none
 version: 1.7.5+480
 
 environment:
-  sdk: ">=2.6.0 <3.0.0"
+  sdk: ">=2.17.1 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  http: ^0.13.3
-  protobuf: ^2.0.0
-  uuid: ^3.0.4
-  package_info_plus: ^1.0.1
-  recase: ^3.0.0
-  share: ^0.6.5+4
+  http: ^1.4.0
+  protobuf: ^4.1.0
+  uuid: ^4.5.1
+  package_info_plus: ^8.3.0
+  recase: ^4.1.0
+  # share: ^0.6.5+4
   flutter_reorderable_list: ^1.2.0
   url_launcher: ^6.0.3
   animated_list_plus: ^0.5.0
-  flutter_keyboard_visibility: ^2.0.0
-  aeyrium_sensor: ^1.0.7
-  sensors: '>=0.4.1+10 <2.0.0'
-  dart_midi: 1.0.1
+  flutter_keyboard_visibility: ^6.0.0
+  # aeyrium_sensor: ^1.0.7
+  # sensors: ^2.0.3
+  # dart_midi: 1.0.1
   quiver: ^3.0.1
-  path_drawing: 0.4.1
+  path_drawing: 1.0.1
   path_provider: ^2.0.2
   shared_preferences: ^2.0.6
-  base_x: ^1.0.1
-  archive: ^3.1.2
-  font_awesome_flutter: ^8.11.0
-  cupertino_icons: ^0.1.2
-  fluro: "^2.0.3"
-  flutter_appavailability: ^0.0.21
-  native_device_orientation: ^0.4.3
-  webview_flutter: ^3.0.4
-  flutter_icons: ^1.1.0
+  base_x: ^2.0.0
+  archive: ^4.0.7
+  font_awesome_flutter: ^10.5.0
+  cupertino_icons: ^1.0.5
+  fluro: ^2.0.3
+  flutter_feather_icons: ^2.0.0
+
+  # flutter_appavailability: ^0.0.21
+  native_device_orientation: ^2.0.3
+  webview_flutter: ^4.13.0
+  # flutter_icons: ^1.1.0
   # flutter_blue: ^0.8.0
-  #  path: ../flutter_blue
   # flutter_reactive_ble: #^3.1.0
-  #   path: ../flutter_reactive_ble
-  flutter_midi_command: #^0.3.0
-   path: ../FlutterMidiCommand
+  flutter_midi_command: ^0.5.3
   dcache: ^0.4.0
   provider: ^6.0.2
-  matrix4_transform: ^2.0.1
+  matrix4_transform: ^4.0.1
+  share_plus: ^11.0.0
+  collection: ^1.19.1
+  appcheck: ^1.5.4+1
+  material_design_icons_flutter: ^7.0.7296
+  sensors_plus: ^6.1.1
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
 
 # The following section is specific to Flutter.
 flutter:
-
   # The following line ensures that the Material Icons font is
   # included with your application, so that you can use the icons in
   # the material Icons class.
