@@ -17,26 +17,25 @@ enum Notehead { quarter, half, whole, percussion }
 
 class _NoteheadInstruction {
   final Notehead notehead;
-  final NoteSign noteSign;
+  final NoteSign? noteSign;
   final double noteheadTop;
   final double noteheadLeft;
   final bool staggered;
-  final bool hadStaggeredNotes;
+  final bool hadStaggeredNotes = false;
   final List<double> ledgerLines;
   const _NoteheadInstruction(
-      {this.notehead,
-      this.noteSign,
-      this.noteheadLeft,
-      this.ledgerLines,
-      this.noteheadTop,
-      this.staggered,
-      this.hadStaggeredNotes});
+      {required this.notehead,
+      required this.noteSign,
+      required this.noteheadLeft,
+      required this.ledgerLines,
+      required this.noteheadTop,
+      required this.staggered});
 }
 
 class _StemInstruction {
   final Offset top;
   final Offset bottom;
-  const _StemInstruction({this.top, this.bottom});
+  const _StemInstruction({required this.top, required this.bottom});
 }
 
 class _RenderInstructions {
@@ -108,7 +107,7 @@ class NotationMusicRenderer extends BaseMusicRenderer {
 
   _renderNoteheadSignAndLedgers(
       Canvas canvas, _NoteheadInstruction instruction) {
-    NoteSign noteSign = instruction.noteSign;
+    NoteSign? noteSign = instruction.noteSign;
     bool staggered = instruction.staggered;
     double noteheadWidth = 18 * xScale;
     double noteheadHeight = noteheadWidth;
@@ -138,46 +137,44 @@ class NotationMusicRenderer extends BaseMusicRenderer {
           Offset(right + 3 * xScale, position * xScale), alphaDrawerPaint);
     });
 
-    if (noteSign != null) {
-      double signLeft, signRight, signTop, signBottom;
-      if (staggered) {
-        signLeft = left -
-            0.7 *
-                noteheadWidth; //bounds.left - 0.7 * noteheadWidth + xOffset;// bounds.right - 2.3 * noteheadWidth;
-      } else {
-        signLeft = left -
-            0.5 *
-                noteheadWidth; //bounds.left - 0.5 * noteheadWidth + xOffset; //bounds.right - 2.8 * noteheadWidth;
-      }
-      signRight = signLeft + 0.5 * noteheadWidth;
-
-      switch (noteSign) {
-        case NoteSign.flat:
-        case NoteSign.double_flat:
-          double difference = noteheadHeight * 1.5;
-          signTop = top - 2 * noteheadHeight / 3 - difference;
-          signBottom = bottom - difference;
-          break;
-        case NoteSign.double_sharp:
-          signTop = top + noteheadHeight / 4;
-          signBottom = bottom - noteheadHeight / 4;
-          break;
-        case NoteSign.sharp:
-        case NoteSign.natural:
-        default:
-          double difference = noteheadHeight * 1.8;
-          signTop = top - difference; //- noteheadHeight / 3;
-          signBottom = bottom - difference; // + noteheadHeight / 3;
-      }
-      final signTopOffset = 26 * yScale;
-      final signLeftOffset = -2 * xScale;
-      signTop += signTopOffset;
-      signBottom += signTopOffset;
-      signLeft += signLeftOffset;
-      signRight += signLeftOffset;
-      Rect signRect = Rect.fromLTRB(signLeft, signTop, signRight, signBottom);
-      _renderSign(canvas, signRect, noteSign);
+    double signLeft, signRight, signTop, signBottom;
+    if (staggered) {
+      signLeft = left -
+          0.7 *
+              noteheadWidth; //bounds.left - 0.7 * noteheadWidth + xOffset;// bounds.right - 2.3 * noteheadWidth;
+    } else {
+      signLeft = left -
+          0.5 *
+              noteheadWidth; //bounds.left - 0.5 * noteheadWidth + xOffset; //bounds.right - 2.8 * noteheadWidth;
     }
+    signRight = signLeft + 0.5 * noteheadWidth;
+
+    switch (noteSign) {
+      case NoteSign.flat:
+      case NoteSign.double_flat:
+        double difference = noteheadHeight * 1.5;
+        signTop = top - 2 * noteheadHeight / 3 - difference;
+        signBottom = bottom - difference;
+        break;
+      case NoteSign.double_sharp:
+        signTop = top + noteheadHeight / 4;
+        signBottom = bottom - noteheadHeight / 4;
+        break;
+      case NoteSign.sharp:
+      case NoteSign.natural:
+      default:
+        double difference = noteheadHeight * 1.8;
+        signTop = top - difference; //- noteheadHeight / 3;
+        signBottom = bottom - difference; // + noteheadHeight / 3;
+    }
+    final signTopOffset = 26 * yScale;
+    final signLeftOffset = -2 * xScale;
+    signTop += signTopOffset;
+    signBottom += signTopOffset;
+    signLeft += signLeftOffset;
+    signRight += signLeftOffset;
+    Rect signRect = Rect.fromLTRB(signLeft, signTop, signRight, signBottom);
+    _renderSign(canvas, signRect, noteSign);
   }
 
   _planNoteheadsSignsAndLedgers(_RenderInstructions result) {
@@ -225,7 +222,7 @@ class NotationMusicRenderer extends BaseMusicRenderer {
         }
 
         // Draw signs
-        NoteSign previousSign;
+        NoteSign? previousSign;
         Iterable<NoteSign> previousSigns = getMostRecentSignsOf(
           note: note,
           relevantMelodies: otherMelodiesOnStaff.followedBy([melody]),
@@ -238,7 +235,7 @@ class NotationMusicRenderer extends BaseMusicRenderer {
           forceShowNaturals = true;
         }
 //        val previousSign = previousSignOf(melody, harmony, note, elementPosition)
-        NoteSign signToDraw;
+        NoteSign? signToDraw;
         switch (note.sign) {
           case NoteSign.sharp:
             if (previousSign != NoteSign.sharp) signToDraw = NoteSign.sharp;
@@ -256,8 +253,7 @@ class NotationMusicRenderer extends BaseMusicRenderer {
             break;
           case NoteSign.natural:
           default:
-            if (forceShowNaturals ||
-                (previousSign != null && previousSign != NoteSign.natural))
+            if (forceShowNaturals || (previousSign != NoteSign.natural))
               signToDraw = NoteSign.natural;
             break;
         }
@@ -307,8 +303,8 @@ class NotationMusicRenderer extends BaseMusicRenderer {
 
   static Map<ArgumentList, Iterable<NoteSign>> recentSignCache = Map();
   Iterable<NoteSign> getMostRecentSignsOf({
-    NoteSpecification note,
-    Iterable<Melody> relevantMelodies,
+    required NoteSpecification note,
+    required Iterable<Melody> relevantMelodies,
   }) {
     final args = ArgumentList([
       relevantMelodies.map((e) => e.id).join(),
@@ -327,8 +323,8 @@ class NotationMusicRenderer extends BaseMusicRenderer {
   }
 
   Iterable<NoteSign> _calculateMostRecentSignsOf({
-    NoteSpecification note,
-    Iterable<Melody> relevantMelodies,
+    required NoteSpecification note,
+    required Iterable<Melody> relevantMelodies,
   }) {
     // final currentBeatPosition = elementPosition.toDouble() / melody.subdivisionsPerBeat;
     // final lastDownbeat = range(0,1000000).firstWhere((it) {
@@ -352,7 +348,7 @@ class NotationMusicRenderer extends BaseMusicRenderer {
         }
         relevantMelodyIndex -= 1;
       }
-      NoteSign melodyResult;
+      NoteSign? melodyResult;
       while (melodyResult == null) {
         if (relevantMelodyIndex % relevantMelody.subdivisionsPerBeat ==
                 0 && // Beginning of measure, stop searching
@@ -394,8 +390,8 @@ class NotationMusicRenderer extends BaseMusicRenderer {
       return [];
     }
     double highestKey =
-        resultCandidates.keys.toList().maxBy((it) => (it * 1000).toInt());
-    return resultCandidates[highestKey];
+        resultCandidates.keys.toList().maxBy((it) => (it * 1000).toInt()) ?? -1;
+    return resultCandidates[highestKey] ?? [];
   }
 
   static Path _sharpPath = parseSvgPathData(
@@ -412,12 +408,12 @@ class NotationMusicRenderer extends BaseMusicRenderer {
         Offset.zero);
   static Path _naturalPath = parseSvgPathData(
       "M 26.578125,106.17187 L 22.640625,107.57812 L 22.640625,75.375001 L 0,85.218751 L 0,1.6875 L 3.796875,1.4210855e-014 L 3.796875,32.765625 L 26.578125,22.359375 L 26.578125,106.17187 z M 22.640625,61.171871 L 22.640625,38.671875 L 3.796875,46.96875 L 3.796875,69.468751 L 22.640625,61.171871 z ");
-  _renderSign(Canvas canvas, Rect signRect, NoteSign sign) {
+  _renderSign(Canvas canvas, Rect signRect, NoteSign? sign) {
 //    canvas.drawRect(signRect, Paint()..style=PaintingStyle.fill..color=Colors.black26);
 //    print("Rendering sign: $sign");
     canvas.save();
     canvas.translate(signRect.topLeft.dx, signRect.topLeft.dy);
-    Path signPath;
+    Path? signPath;
     switch (sign) {
       case NoteSign.sharp:
         signPath = _sharpPath;
@@ -449,6 +445,7 @@ class NotationMusicRenderer extends BaseMusicRenderer {
       default:
         break;
     }
+
     if (signPath != null) {
       canvas.drawPath(signPath, alphaDrawerPaint);
     }
