@@ -1,36 +1,31 @@
 import 'package:beatscratch_flutter_redux/util/music_theory.dart';
-import 'package:beatscratch_flutter_redux/util/ui_utils.dart';
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:font_awesome_flutter/font_awesome_flutter.dart';
-import 'package:package_info_plus/package_info_plus.dart';
 
-import 'beatscratch_plugin.dart';
 import 'colors.dart';
-import 'export/export_ui.dart';
 import 'generated/protos/protos.dart';
 import 'music_preview/melody_preview.dart';
 import 'music_preview/part_preview.dart';
 import 'music_preview/section_preview.dart';
 import 'ui_models.dart';
 import 'widget/beats_badge.dart';
-import 'widget/my_platform.dart';
 
 Future<Object> showEditMenu(
-    {@required BuildContext context,
-    @required RelativeRect position,
-    @required Score score,
-    @required Part part,
-    @required Melody selectedMelody,
-    @required MusicViewMode musicViewMode,
-    @required Section section,
-    @required Function(Object) editObject}) async {
+    {required BuildContext context,
+    required RelativeRect position,
+    required Score score,
+    required Part? part,
+    Melody? selectedMelody,
+    required MusicViewMode musicViewMode,
+    required Section section,
+    required Function(Object) editObject}) async {
   onSelected(Object object) {
     Navigator.pop(context);
     editObject(object);
   }
 
-  final melodies = part.melodies.where((m) => section.referenceTo(m).isEnabled);
+  final melodies =
+      part?.melodies.where((m) => section.referenceTo(m)?.isEnabled == true) ??
+          [];
   return showMenu(
       context: context,
       position: position,
@@ -38,6 +33,7 @@ Future<Object> showEditMenu(
       color: (musicBackgroundColor.luminance < 0.5
               ? subBackgroundColor
               : musicBackgroundColor)
+          // .withAlpha(240)
           .withOpacity(0.95),
       items: [
         PopupMenuItem(
@@ -81,7 +77,7 @@ Future<Object> showEditMenu(
             value: null,
             child: Column(children: [
               Row(children: [
-                Text(part.midiName,
+                Text(part?.midiName ?? '',
                     textAlign: TextAlign.left,
                     // overflow: TextOverflow.ellipsis,
                     softWrap: true,
@@ -103,7 +99,7 @@ Future<Object> showEditMenu(
             enabled: false,
           ),
         ...melodies.map((m) => melodyMenuItem(
-            score, m, part, section, onSelected, m == selectedMelody)),
+            score, m, part!, section, onSelected, m == selectedMelody)),
       ]);
 }
 
@@ -195,13 +191,14 @@ Material sectionPreview(Score score, Section section,
   );
 }
 
-PopupMenuItem<Part> partMenuItem(Score score, Part part, Section section,
+PopupMenuItem<Part> partMenuItem(Score score, Part? part, Section section,
     Function(Object) onSelected, bool isSelected) {
   return PopupMenuItem(
       value: part,
       mouseCursor: SystemMouseCursors.basic,
       padding: EdgeInsets.zero,
-      child: partPreview(score, part, section, onSelected, isSelected),
+      child:
+          partPreview(score, part ?? Part(), section, onSelected, isSelected),
       enabled: true);
 }
 
