@@ -1,21 +1,23 @@
 import 'dart:io';
 
-import '../widget/my_platform.dart';
-
+import 'package:beatscratch_flutter_redux/generated/protos/protos.dart';
+import 'package:beatscratch_flutter_redux/util/music_theory.dart';
 import 'package:path_provider/path_provider.dart';
 
-import '../util/music_theory.dart';
+import '../widget/my_platform.dart';
 import 'export_models.dart';
 
 class ExportManager {
-  Directory exportsDirectory;
+  late Directory exportsDirectory;
 
   File createExportFile(BSExport export) {
     String path =
         "${exportsDirectory.path.toString()}/${Uri.encodeComponent(export.score.name).replaceAll("%20", " ")}";
-    if (export.exportedSection != null) {
-      path += "-${Uri.encodeComponent(export.exportedSection.canonicalName)}";
+    Section? exportedSection = export.exportedSection;
+    if (exportedSection != null) {
+      path += "-${Uri.encodeComponent(exportedSection.canonicalName)}";
     }
+
     path +=
         "-${DateTime.now().toString().replaceAll(" ", '-').replaceAll(":", '-').split(".").first}";
     path += ".${export.exportType.toString().split(".").last}";
@@ -25,13 +27,10 @@ class ExportManager {
   }
 
   List<FileSystemEntity> get exportFiles {
-    if (exportsDirectory != null) {
-      List<FileSystemEntity> result = exportsDirectory?.listSync();
-      result.sort(
-          (a, b) => b.statSync().modified.compareTo(a.statSync().modified));
-      return result;
-    }
-    return [];
+    List<FileSystemEntity> result = exportsDirectory.listSync();
+    result
+        .sort((a, b) => b.statSync().modified.compareTo(a.statSync().modified));
+    return result;
   }
 
   ExportManager() {
