@@ -1,9 +1,11 @@
 import 'dart:io';
+import 'package:share_plus/share_plus.dart';
+
 import '../beatscratch_plugin.dart';
 import '../generated/protos/music.pb.dart';
 import '../messages/messages_ui.dart';
 import '../util/dummydata.dart';
-import 'package:share/share.dart';
+// import 'package:share/share.dart';
 
 import '../util/util.dart';
 
@@ -24,7 +26,7 @@ class ExportUI {
   static final double _baseHeight = 220;
   static final double _progressHeight = 30;
 
-  static Widget exportIcon({double size = 24, Color color}) =>
+  static Widget exportIcon({double size = 24, Color? color}) =>
       Transform.translate(
         offset: Offset(size * 2 / 24, 0),
         child: Icon(MyPlatform.isAppleOS ? CupertinoIcons.share : Icons.share,
@@ -33,9 +35,9 @@ class ExportUI {
 
   bool visible = false;
   bool exporting = false;
-  final BSExport export = BSExport(score: defaultScore());
+  final BSExport export = BSExport(defaultScore());
   ExportManager exportManager = ExportManager();
-  MessagesUI messagesUI;
+  MessagesUI? messagesUI;
 
   double get baseHeight => visible ? _baseHeight : 0.0;
 
@@ -44,9 +46,9 @@ class ExportUI {
   double get height => baseHeight + progressHeight;
 
   Widget build(
-      {@required BuildContext context,
-      @required Function(VoidCallback) setState,
-      @required Section currentSection}) {
+      {required BuildContext context,
+      required Function(VoidCallback) setState,
+      required Section currentSection}) {
     return AnimatedContainer(
       duration: animationDuration,
       height: height,
@@ -104,7 +106,7 @@ class ExportUI {
                           Expanded(child: SizedBox()),
                           Transform.translate(
                             offset: Offset(0, 3),
-                            child: Text(export?.score?.name ?? "null",
+                            child: Text(export.score.name ?? "null",
                                 maxLines: 1,
                                 overflow: TextOverflow.ellipsis,
                                 style: TextStyle(
@@ -193,7 +195,7 @@ class ExportUI {
                                   exporting = true;
                                   visible = false;
                                   Future.microtask(() {
-                                    File file;
+                                    File? file;
                                     bool success = false;
                                     try {
                                       file = export(exportManager);
@@ -204,7 +206,7 @@ class ExportUI {
                                         print(e.stackTrace);
                                       }
 
-                                      messagesUI.sendMessage(
+                                      messagesUI?.sendMessage(
                                           message: "MIDI Export failed!",
                                           isError: true);
                                     }
@@ -214,35 +216,45 @@ class ExportUI {
                                       });
                                       if (success) {
                                         if (MyPlatform.isMacOS) {
-                                          messagesUI.sendMessage(
+                                          messagesUI?.sendMessage(
                                               message:
                                                   "Opening exports directory in Finder...");
                                           Future.delayed(Duration(seconds: 1),
                                               () {
                                             launchURL(
                                                 "file://${exportManager.exportsDirectory.path}");
-                                            messagesUI.sendMessage(
+                                            messagesUI?.sendMessage(
                                                 message: "Export complete!");
                                           });
                                         } else if (MyPlatform.isIOS) {
-                                          messagesUI.sendMessage(
+                                          messagesUI?.sendMessage(
                                               message:
                                                   "Opening exports directory in Files...");
                                           Future.delayed(Duration(seconds: 1),
                                               () {
                                             launchURL(
                                                 "shareddocuments://${exportManager.exportsDirectory.path}");
-                                            messagesUI.sendMessage(
+                                            messagesUI?.sendMessage(
                                                 message: "Export complete!");
                                           });
                                         } else if (MyPlatform.isMobile) {
-                                          messagesUI.sendMessage(
+                                          messagesUI?.sendMessage(
                                               message: "Sharing MIDI file...");
                                           Future.delayed(Duration(seconds: 1),
                                               () {
-                                            Share.shareFiles([file.path],
-                                                text: export.score.name);
-                                            messagesUI.sendMessage(
+                                            SharePlus.instance
+                                                .share(ShareParams(
+                                              files: [
+                                                XFile(file!.path),
+                                              ],
+                                              // sharePositionOrigin: box!.localToGlobal(Offset.zero) & box.size,
+                                              // fileNameOverrides: [fileName],
+                                              downloadFallbackEnabled: true,
+                                            ));
+
+                                            // [file.path],
+                                            // text: export.score.name);
+                                            messagesUI?.sendMessage(
                                                 message: "Export complete!");
                                           });
                                         }
@@ -269,9 +281,9 @@ class ExportUI {
       EdgeInsets.only(left: 5, top: 5, bottom: 5);
 
   SingleChildScrollView exportOptions(
-      {@required BuildContext context,
-      @required Function(VoidCallback) setState,
-      @required Section currentSection}) {
+      {required BuildContext context,
+      required Function(VoidCallback) setState,
+      required Section currentSection}) {
     double width = MediaQuery.of(context).size.width;
     double scrollContainerWidth = width - 44;
     double exportTypeWidth = 100;
@@ -282,18 +294,15 @@ class ExportUI {
         exportTypeWidth + sectionWidth + partsWidth + speedWidth;
     bool usesFlex = scrollContainerWidth > scrollContentsWidth;
 
-    Widget exportOption(String label, String value, double size, Color color,
-        {List<String> values,
-        List<String> disabledValues,
-        Widget customValue,
-        Function(String) selectValue}) {
-      if (disabledValues == null) {
-        disabledValues = [];
-      }
-      if (values == null || values.isEmpty) {
+    Widget exportOption(String label, String? value, double size, Color color,
+        {required List<String?> values,
+        List<String> disabledValues = const [],
+        Widget? customValue,
+        required Function(String?) selectValue}) {
+      if (values.isEmpty) {
         values = [value];
       }
-      if (value != null && !values.contains(value)) {
+      if (!values.contains(value)) {
         values = values + [value];
       }
       values = values.toSet().toList();
@@ -301,12 +310,12 @@ class ExportUI {
       Widget column = Column(children: [
         // if (usesFlex) Expanded(child:SizedBox()),
         ...values.map((v) {
-          final clickable = selectValue != null && !disabledValues.contains(v);
+          final clickable = !disabledValues.contains(v);
           return MyFlatButton(
               color: v == value ? Colors.white : Colors.black12,
               padding: EdgeInsets.all(5),
               onPressed: clickable ? () => selectValue(v) : null,
-              child: Text(v,
+              child: Text(v ?? "",
                   textAlign: TextAlign.center,
                   style: valueStyle(
                     !clickable
@@ -316,7 +325,7 @@ class ExportUI {
                             : Colors.white,
                   )));
         }).toList(),
-        if (customValue != null) customValue,
+        customValue ?? SizedBox(),
         // if (usesFlex) Expanded(child:SizedBox())
       ]);
       // if (!usesFlex) {
@@ -341,8 +350,7 @@ class ExportUI {
       }
     }
 
-    if (export.score == null ||
-        !export.score.sections.any((s) => s.id == export.sectionId)) {
+    if (!export.score.sections.any((s) => s.id == export.sectionId)) {
       export.sectionId = null;
     }
     export.partIds.removeWhere(
@@ -370,7 +378,8 @@ class ExportUI {
             "${(BeatScratchPlugin.bpmMultiplier ?? 1).toStringAsFixed(3).replaceAll("1.000", "1")}x"
           ],
           selectValue: (v) => setState(() {
-                export.tempoMultiplier = double.parse(v.replaceAll("x", ""));
+                export.tempoMultiplier =
+                    double.parse((v ?? "").replaceAll("x", ""));
               })),
       exportOption(
           "Section",
@@ -378,7 +387,7 @@ class ExportUI {
               ? "Entire Score"
               : export.score.sections
                       .firstWhere((s) => s.id == export.sectionId)
-                      ?.canonicalName ??
+                      .canonicalName ??
                   "NULL",
           sectionWidth,
           chromaticSteps[3],
