@@ -1,8 +1,6 @@
 import 'dart:math';
 
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/rendering.dart';
 import 'package:flutter_reorderable_list/flutter_reorderable_list.dart' as frl;
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:animated_list_plus/animated_list_plus.dart';
@@ -29,7 +27,7 @@ class LayersPartView extends StatefulWidget {
       MelodyReferenceView.columnWidthMicroIncrement;
 
   final Score score;
-  final Axis scrollDirection;
+  final Axis? scrollDirection;
   final Function(Melody) selectMelody;
   final VoidCallback toggleEditingMelody;
   final VoidCallback hideMelodyView;
@@ -39,8 +37,8 @@ class LayersPartView extends StatefulWidget {
   final Function(Part) selectPart;
   final Color sectionColor;
   final Section currentSection;
-  final Melody selectedMelody;
-  final Part selectedPart;
+  final Melody? selectedMelody;
+  final Part? selectedPart;
   final Part colorboardPart;
   final Part keyboardPart;
   final Function(Part) setKeyboardPart;
@@ -60,33 +58,33 @@ class LayersPartView extends StatefulWidget {
   }
 
   LayersPartView({
-    this.score,
+    required this.score,
     this.scrollDirection,
-    this.part,
-    this.sectionColor,
-    this.currentSection,
-    this.selectedMelody,
-    this.selectMelody,
-    this.colorboardPart,
-    this.keyboardPart,
-    this.setKeyboardPart,
-    this.setColorboardPart,
-    this.selectPart,
-    this.toggleMelodyReference,
-    this.setReferenceVolume,
-    this.setPartVolume,
-    this.editingMelody,
-    this.toggleEditingMelody,
-    this.hideMelodyView,
-    this.removePart,
-    this.selectedPart,
-    this.enableColorboard,
-    this.showBeatCounts,
-    this.height,
-    this.showMediumDetails,
-    this.autoScroll,
-    this.showHighDetails,
-    this.width,
+    required this.part,
+    required this.sectionColor,
+    required this.currentSection,
+    required this.selectedMelody,
+    required this.selectMelody,
+    required this.colorboardPart,
+    required this.keyboardPart,
+    required this.setKeyboardPart,
+    required this.setColorboardPart,
+    required this.selectPart,
+    required this.toggleMelodyReference,
+    required this.setReferenceVolume,
+    required this.setPartVolume,
+    required this.editingMelody,
+    required this.toggleEditingMelody,
+    required this.hideMelodyView,
+    required this.removePart,
+    required this.selectedPart,
+    required this.enableColorboard,
+    required this.showBeatCounts,
+    required this.height,
+    required this.showMediumDetails,
+    required this.autoScroll,
+    required this.showHighDetails,
+    required this.width,
   });
 
   @override
@@ -122,7 +120,7 @@ class _LayersPartViewState extends State<LayersPartView> {
 
   Section get currentSection => widget.currentSection;
 
-  Melody get selectedMelody => widget.selectedMelody;
+  Melody? get selectedMelody => widget.selectedMelody;
 
   get setReferenceVolume => widget.setReferenceVolume;
 
@@ -143,7 +141,7 @@ class _LayersPartViewState extends State<LayersPartView> {
   get toggleEditingMelody => widget.toggleEditingMelody;
 
   get hideMelodyView => widget.hideMelodyView;
-  ScrollController scrollController;
+  late ScrollController scrollController;
 
   int _indexOfKey(Key key) {
     return widget._items.indexWhere((Melody melody) => Key(melody.id) == key);
@@ -174,7 +172,7 @@ class _LayersPartViewState extends State<LayersPartView> {
   bool get scrollControllerIsNearTop =>
       scrollController.hasClients &&
       scrollController.offset < (widget.showMediumDetails ? 150 : 165);
-  DateTime _lastScrollTime;
+  DateTime? _lastScrollTime;
   setScrollToTopTimeout() {
     final v = DateTime.now();
     Future.delayed(Duration(seconds: 3), () {
@@ -267,8 +265,8 @@ class _LayersPartViewState extends State<LayersPartView> {
       ];
   int newMelodyBeatCountIndex = 0;
 
-  String lastSelectedMelodyId;
-  double prevWidth;
+  String? lastSelectedMelodyId;
+  double? prevWidth;
 
   @override
   Widget build(BuildContext context) {
@@ -288,9 +286,8 @@ class _LayersPartViewState extends State<LayersPartView> {
       backgroundColor = isSelectedPart ? Colors.white : Colors.grey;
       textColor = isSelectedPart ? Colors.grey : Colors.white;
     }
-    if (lastSelectedMelodyId != null &&
-        selectedMelody != null &&
-        lastSelectedMelodyId != selectedMelody.id &&
+    if (selectedMelody != null &&
+        lastSelectedMelodyId != selectedMelody?.id &&
         widget.autoScroll) {
       int indexOfMelody =
           part.melodies.indexWhere((m) => m.id == selectedMelody?.id);
@@ -587,7 +584,7 @@ class _LayersPartViewState extends State<LayersPartView> {
       expandedHeight: 50.0,
       flexibleSpace: MyFlatButton(
           onPressed: () {
-            final melody = widget.selectedMelody.bsCopy()..id = uuid.v4();
+            final melody = widget.selectedMelody!.bsCopy()..id = uuid.v4();
             bool melodyExists = melody.name.isNotEmpty &&
                 part.melodies.any((it) => it.name == melody.name);
             if (melodyExists && melody.name.isNotEmpty) {
@@ -595,9 +592,9 @@ class _LayersPartViewState extends State<LayersPartView> {
                 r"^(.*?)(\d*)\s*$",
               );
               final match = expr.allMatches(melody.name).first;
-              String prefix = match.group(1);
+              String prefix = match.group(1)!;
               prefix = prefix.trim();
-              int number = int.tryParse(match.group(2)) ?? 1;
+              int number = int.tryParse(match.group(2)!) ?? 1;
               int newNumber = number + 1;
               melody.name = "$prefix $newNumber";
               while (part.melodies.any((it) {
@@ -622,11 +619,11 @@ class _LayersPartViewState extends State<LayersPartView> {
                       child: Transform.translate(
                           offset: Offset(5, -1),
                           child: Text(
-                            selectedMelody.canonicalName,
+                            selectedMelody?.canonicalName ?? '',
                             maxLines: 1,
                             overflow: TextOverflow.fade,
                             style: TextStyle(
-                                color: selectedMelody.name?.isNotEmpty ?? false
+                                color: selectedMelody?.name.isNotEmpty ?? false
                                     ? textColor
                                     : textColor.withOpacity(0.5),
                                 fontWeight: FontWeight.w300),
@@ -638,7 +635,7 @@ class _LayersPartViewState extends State<LayersPartView> {
                 child: Transform.translate(
                   offset: Offset(5, 0),
                   child: BeatsBadge(
-                    beats: selectedMelody.beatCount,
+                    beats: selectedMelody?.beatCount ?? 0,
                     show: widget.showBeatCounts,
                   ),
                 )),
