@@ -1,8 +1,6 @@
 import 'dart:math';
 
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/rendering.dart';
 import 'package:animated_list_plus/animated_list_plus.dart';
 import 'package:animated_list_plus/transitions.dart';
 
@@ -29,7 +27,7 @@ class LayersView extends StatefulWidget {
   final Score score;
   final Color sectionColor;
   final Section currentSection;
-  final Melody selectedMelody;
+  final Melody? selectedMelody;
   final Function(Melody) selectMelody;
   final VoidCallback toggleEditingMelody;
   final VoidCallback hideMelodyView;
@@ -38,7 +36,7 @@ class LayersView extends StatefulWidget {
   final Function(Part, double) setPartVolume;
   final Part colorboardPart;
   final Part keyboardPart;
-  final Part selectedPart;
+  final Part? selectedPart;
   final Function(Part) setKeyboardPart;
   final Function(Part) setColorboardPart;
   final Function(Part) selectPart;
@@ -51,34 +49,34 @@ class LayersView extends StatefulWidget {
   final bool shiftUpZoomControls;
 
   LayersView(
-      {this.musicViewMode,
-      this.superSetState,
-      this.appSettings,
-      this.score,
-      this.sectionColor,
-      this.currentSection,
-      this.selectedMelody,
-      this.selectMelody,
-      this.colorboardPart,
-      this.keyboardPart,
-      this.setKeyboardPart,
-      this.setColorboardPart,
-      this.selectPart,
-      this.toggleMelodyReference,
-      this.setReferenceVolume,
-      this.setPartVolume,
-      this.editingMelody,
-      this.toggleEditingMelody,
-      this.hideMelodyView,
-      this.availableWidth,
-      this.selectedPart,
-      this.enableColorboard,
-      this.showBeatCounts,
-      this.height,
-      Key key,
-      this.showViewOptions,
-      this.scoreManager,
-      this.shiftUpZoomControls})
+      {Key? key,
+      required this.musicViewMode,
+      required this.superSetState,
+      required this.appSettings,
+      required this.score,
+      required this.sectionColor,
+      required this.currentSection,
+      required this.selectedMelody,
+      required this.selectMelody,
+      required this.colorboardPart,
+      required this.keyboardPart,
+      required this.setKeyboardPart,
+      required this.setColorboardPart,
+      required this.selectPart,
+      required this.toggleMelodyReference,
+      required this.setReferenceVolume,
+      required this.setPartVolume,
+      required this.editingMelody,
+      required this.toggleEditingMelody,
+      required this.hideMelodyView,
+      required this.availableWidth,
+      required this.selectedPart,
+      required this.enableColorboard,
+      required this.showBeatCounts,
+      required this.height,
+      required this.showViewOptions,
+      required this.scoreManager,
+      required this.shiftUpZoomControls})
       : super(key: key);
 
   @override
@@ -144,10 +142,6 @@ class _LayersViewState extends State<LayersView> {
                             Part part = newDrumPart();
                             widget.score.parts.add(part);
                             BeatScratchPlugin.createPart(part);
-//                          BeatScratchPlugin.pushPart(part);
-                            if (widget.keyboardPart == null) {
-                              widget.setKeyboardPart(part);
-                            }
                             widget.selectPart(part);
                           });
                         });
@@ -200,12 +194,6 @@ class _LayersViewState extends State<LayersView> {
                           widget.score.parts.add(part);
                         }
                         BeatScratchPlugin.createPart(part);
-                        if (widget.keyboardPart == null) {
-                          widget.setKeyboardPart(part);
-                        }
-                        if (widget.colorboardPart == null) {
-                          widget.setColorboardPart(part);
-                        }
                         widget.selectPart(part);
                       });
                     });
