@@ -1,6 +1,8 @@
+import 'dart:async';
 import 'dart:math';
 
 import 'package:beatscratch_flutter_redux/storage/universe_manager.dart';
+import 'package:collection/collection.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:url_launcher/url_launcher.dart';
 
@@ -9,9 +11,7 @@ import 'universe_view/universe_view.dart';
 
 import 'recording/recording.dart';
 import 'package:fluro/fluro.dart' as Fluro;
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/rendering.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_keyboard_visibility/flutter_keyboard_visibility.dart';
 import 'package:native_device_orientation/native_device_orientation.dart';
@@ -32,7 +32,6 @@ import 'storage/score_manager.dart';
 import 'storage/score_picker.dart';
 import 'storage/url_conversions.dart';
 import 'ui_models.dart';
-import 'util/bs_methods.dart';
 import 'util/dummydata.dart';
 import 'util/music_theory.dart';
 import 'util/proto_utils.dart';
@@ -67,17 +66,15 @@ const Map<int, Color> swatch = {
 ScoreManager _scoreManager = ScoreManager();
 UniverseManager _universeManager = UniverseManager();
 AppSettings _appSettings = AppSettings();
-var baseHandler = Fluro.Handler(
-    handlerFunc: (BuildContext context, Map<String, dynamic> params) {
+var baseHandler = Fluro.Handler(handlerFunc: (context, params) {
   return MyHomePage(title: 'BeatScratch', initialScore: defaultScore());
   // return UsersScreen(params["scoreData"][0]);
 });
-var scoreRouteHandler = Fluro.Handler(
-    handlerFunc: (BuildContext context, Map<String, dynamic> params) {
-  String scoreData = params["scoreData"][0];
+var scoreRouteHandler = Fluro.Handler(handlerFunc: (context, params) {
+  String scoreData = params["scoreData"]![0];
   Score score;
   try {
-    score = scoreFromUrlHashValue(scoreData);
+    score = scoreFromUrlHashValue(scoreData)!;
   } catch (any) {
     score = defaultScore();
   }
@@ -85,9 +82,8 @@ var scoreRouteHandler = Fluro.Handler(
   return MyHomePage(title: 'BeatScratch', initialScore: score);
   // return UsersScreen(params["scoreData"][0]);
 });
-var pastebinRouteHandler = Fluro.Handler(
-    handlerFunc: (BuildContext context, Map<String, dynamic> params) {
-  String pastebinCode = params["pasteBinData"][0];
+var pastebinRouteHandler = Fluro.Handler(handlerFunc: (context, params) {
+  String pastebinCode = params["pasteBinData"]![0];
   return MyHomePage(
     title: 'BeatScratch',
     initialScore: defaultScore(),
@@ -111,7 +107,7 @@ class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
 //    debugPaintSizeEnabled = true;
-    MyHomePage home;
+    late MyHomePage home;
     try {
       home = MyHomePage(title: 'BeatScratch', initialScore: defaultScore());
     } catch (e) {
@@ -145,21 +141,25 @@ class MyApp extends StatelessWidget {
 }
 
 class MyHomePage extends StatefulWidget {
-  MyHomePage({Key key, this.title, this.initialScore, this.pastebinCode})
+  MyHomePage(
+      {Key? key,
+      required this.title,
+      required this.initialScore,
+      this.pastebinCode})
       : super(key: key);
 
   final String title;
   final Score initialScore;
-  final String pastebinCode;
+  final String? pastebinCode;
 
   @override
   _MyHomePageState createState() => _MyHomePageState();
 }
 
 class _MyHomePageState extends State<MyHomePage> with TickerProviderStateMixin {
-  Score score;
+  late Score score;
   InteractionMode interactionMode = InteractionMode.view;
-  SplitMode _splitMode;
+  late SplitMode _splitMode;
 
   SplitMode get splitMode => _splitMode;
 
@@ -199,7 +199,7 @@ class _MyHomePageState extends State<MyHomePage> with TickerProviderStateMixin {
   set recordingMelody(value) {
     _recordingMelody = value;
     if (value) {
-      BeatScratchPlugin.setRecordingMelody(selectedMelody);
+      BeatScratchPlugin.setRecordingMelody(value);
       _showMusicView();
     } else {
       BeatScratchPlugin.setRecordingMelody(null);
@@ -209,7 +209,7 @@ class _MyHomePageState extends State<MyHomePage> with TickerProviderStateMixin {
     }
   }
 
-  Section _currentSection; //
+  Section _currentSection = Section();
 
   Section get currentSection => _currentSection;
 
@@ -219,7 +219,7 @@ class _MyHomePageState extends State<MyHomePage> with TickerProviderStateMixin {
         min(section.beatCount - 1, BeatScratchPlugin.currentBeat.value);
     _currentSection = section;
     if (recordingMelody &&
-        section.referenceTo(selectedMelody).playbackType ==
+        section.referenceTo(selectedMelody!).playbackType ==
             MelodyReference_PlaybackType.disabled) {
       recordingMelody = false;
     }
@@ -230,7 +230,7 @@ class _MyHomePageState extends State<MyHomePage> with TickerProviderStateMixin {
     }
   }
 
-  int _tapInBeat;
+  int? _tapInBeat;
   bool showViewOptions = false;
   bool _wasKeyboardShowingWhenMidiConfigurationOpened = false;
   bool _wasColorboardShowingWhenMidiConfigurationOpened = false;
@@ -254,7 +254,7 @@ class _MyHomePageState extends State<MyHomePage> with TickerProviderStateMixin {
   bool _showKeyboardConfiguration = false;
   bool _enableColorboard = false;
 
-  get enableColorboard => _enableColorboard;
+  bool get enableColorboard => _enableColorboard;
 
   set enableColorboard(bool value) {
     _enableColorboard = value;
@@ -264,36 +264,30 @@ class _MyHomePageState extends State<MyHomePage> with TickerProviderStateMixin {
 
   bool showColorboard = false;
   bool _showColorboardConfiguration = false;
-  Part _keyboardPart;
+  Part? _keyboardPart;
 
-  Part get keyboardPart => _keyboardPart;
+  Part? get keyboardPart => _keyboardPart;
 
-  set keyboardPart(Part part) {
+  set keyboardPart(Part? part) {
     _keyboardPart = part;
-    if (part == null) {
-      showKeyboard = false;
-    }
-    BeatScratchPlugin.setKeyboardPart(part);
+    if (part != null) BeatScratchPlugin.setKeyboardPart(part);
   }
 
-  Part _colorboardPart;
+  Part? _colorboardPart;
 
-  Part get colorboardPart => _colorboardPart;
+  Part? get colorboardPart => _colorboardPart;
 
-  set colorboardPart(Part part) {
+  set colorboardPart(Part? part) {
     _colorboardPart = part;
-    if (part == null) {
-      showColorboard = false;
-    }
 //    BeatScratchPlugin.setColorboardPart(part);
   }
 
-  ValueNotifier<Iterable<int>> colorboardNotesNotifier;
-  ValueNotifier<Iterable<int>> keyboardNotesNotifier;
-  ValueNotifier<Map<String, List<int>>> bluetoothControllerPressedNotes;
-  int keyboardChordBase;
-  Set<int> keyboardChordNotes = Set();
-  ValueNotifier<Chord> keyboardChordNotifier;
+  late ValueNotifier<Iterable<int>> colorboardNotesNotifier;
+  late ValueNotifier<Iterable<int>> keyboardNotesNotifier;
+  late ValueNotifier<Map<String, List<int>>> bluetoothControllerPressedNotes;
+  int? keyboardChordBase;
+  final Set<int> keyboardChordNotes = Set();
+  late ValueNotifier<Chord> keyboardChordNotifier;
 
   bool get melodyViewVisible => _musicViewSizeFactor > 0;
 
@@ -326,9 +320,6 @@ class _MyHomePageState extends State<MyHomePage> with TickerProviderStateMixin {
     setState(() {
       _musicViewSizeFactor = 0;
       _prevSelectedMelody = selectedMelody;
-      if (_prevSelectedMelody == null) {
-        _prevSelectedPart = selectedPart;
-      }
       selectedMelody = null;
       recordingMelody = false;
       selectedPart = null;
@@ -358,9 +349,7 @@ class _MyHomePageState extends State<MyHomePage> with TickerProviderStateMixin {
     setState(() {
       bool wasAssignedByPartCreation = keyboardPart == null;
       keyboardPart = part;
-      if (part != null &&
-          !wasAssignedByPartCreation &&
-          !hasPrioritizedMIDIController) {
+      if (!wasAssignedByPartCreation && !hasPrioritizedMIDIController) {
         showKeyboard = true;
       }
     });
@@ -370,56 +359,48 @@ class _MyHomePageState extends State<MyHomePage> with TickerProviderStateMixin {
     setState(() {
       bool wasAssignedByPartCreation = colorboardPart == null;
       colorboardPart = part;
-      if (part != null && !wasAssignedByPartCreation) {
+      if (!wasAssignedByPartCreation) {
         showColorboard = true;
       }
     });
   }
 
-  Melody _selectedMelody;
+  Melody? _selectedMelody;
 
-  Melody get selectedMelody => _selectedMelody;
+  Melody? get selectedMelody => _selectedMelody;
 
-  set selectedMelody(Melody selectedMelody) {
+  set selectedMelody(Melody? selectedMelody) {
     _selectedMelody = selectedMelody;
-    if (selectedMelody != null) {
-      Part part = score.parts
-          .firstWhere((p) => p.melodies.any((m) => m.id == selectedMelody.id));
-      if (part != null) {
-        keyboardPart = part;
-      }
-    } else {
-      recordingMelody = false;
+    Part? part = score.parts.firstWhereOrNull(
+        (p) => p.melodies.any((m) => m.id == selectedMelody?.id));
+    if (part != null) {
+      keyboardPart = part;
     }
   }
 
-  Part _selectedPart;
+  Part? _selectedPart;
 
-  Part get selectedPart => _selectedPart;
+  Part? get selectedPart => _selectedPart;
 
-  set selectedPart(Part selectedPart) {
+  set selectedPart(Part? selectedPart) {
     _selectedPart = selectedPart;
-    if (selectedPart != null) {
-      keyboardPart = selectedPart;
-    }
+    keyboardPart = selectedPart;
   }
 
-  Part _viewingPart;
+  Part? _viewingPart;
 
-  Part get viewingPart => _viewingPart;
+  Part? get viewingPart => _viewingPart;
 
-  set viewingPart(Part viewingPart) {
+  set viewingPart(Part? viewingPart) {
     _viewingPart = viewingPart;
-    if (viewingPart != null) {
-      keyboardPart = viewingPart;
-    }
+    keyboardPart = viewingPart;
   }
 
   List<SectionList> _sectionLists = [];
 
   Color get sectionColor => currentSection.color.color;
 
-  _selectOrDeselectMelody(Melody melody, {bool hideMusicOnDeselect: true}) {
+  _selectOrDeselectMelody(Melody? melody, {bool hideMusicOnDeselect = true}) {
     setState(() {
       if (selectedMelody != melody) {
         selectedMelody = melody;
@@ -435,7 +416,7 @@ class _MyHomePageState extends State<MyHomePage> with TickerProviderStateMixin {
         recordingMelody = false;
         if (hideMusicOnDeselect) {
           _hideMusicView();
-        } else {
+        } else if (melody != null) {
           final part = score.parts
               .firstWhere((p) => p.melodies.any((m) => m.id == melody.id));
           _selectOrDeselectPart(part, hideMusicOnDeselect: hideMusicOnDeselect);
@@ -444,7 +425,7 @@ class _MyHomePageState extends State<MyHomePage> with TickerProviderStateMixin {
     });
   }
 
-  _selectOrDeselectPart(Part part, {bool hideMusicOnDeselect: true}) {
+  _selectOrDeselectPart(Part part, {bool hideMusicOnDeselect = true}) {
     setState(() {
       print("yay");
       if (selectedPart != part) {
@@ -459,8 +440,10 @@ class _MyHomePageState extends State<MyHomePage> with TickerProviderStateMixin {
           _hideMusicView();
         } else {
           if (musicViewMode == MusicViewMode.melody) {
-            _selectOrDeselectMelody(selectedMelody,
-                hideMusicOnDeselect: hideMusicOnDeselect);
+            if (selectedMelody != null) {
+              _selectOrDeselectMelody(selectedMelody!,
+                  hideMusicOnDeselect: hideMusicOnDeselect);
+            }
           } else {
             selectedPart = null;
             _prevSelectedPart = null;
@@ -479,9 +462,7 @@ class _MyHomePageState extends State<MyHomePage> with TickerProviderStateMixin {
     } else {
       setState(() {
         ref.playbackType = MelodyReference_PlaybackType.disabled;
-        if (selectedMelody != null &&
-            ref != null &&
-            ref.melodyId == selectedMelody.id) {
+        if (ref.melodyId == selectedMelody?.id) {
           recordingMelody = false;
         }
       });
@@ -565,8 +546,8 @@ class _MyHomePageState extends State<MyHomePage> with TickerProviderStateMixin {
     });
   }
 
-  Part _prevSelectedPart;
-  Melody _prevSelectedMelody;
+  Part? _prevSelectedPart;
+  Melody? _prevSelectedMelody;
 
   _editMode() {
     BeatScratchPlugin.setPlaybackMode(Playback_Mode.section);
@@ -574,27 +555,9 @@ class _MyHomePageState extends State<MyHomePage> with TickerProviderStateMixin {
       // _universeManager.currentUniverseScore = "";
       universeViewUI.visible = false;
       if (interactionMode.isEdit) {
-        if (selectedMelody != null) {
-          _prevSelectedMelody = selectedMelody;
-          _prevSelectedPart = null;
-          _hideMusicView();
-        } else if (selectedPart != null) {
-          _prevSelectedMelody = null;
-          _prevSelectedPart = selectedPart;
-          _hideMusicView();
-        } else if (musicViewMode == MusicViewMode.section) {
-          _prevSelectedMelody = null;
-          _prevSelectedPart = null;
-          _hideMusicView();
-        } else {
-          if (_prevSelectedMelody != null) {
-            _selectOrDeselectMelody(_prevSelectedMelody);
-          } else if (_prevSelectedPart != null) {
-            _selectOrDeselectPart(_prevSelectedPart);
-          } else {
-            _selectSection(currentSection);
-          }
-        }
+        _prevSelectedMelody = selectedMelody;
+        _prevSelectedPart = null;
+        _hideMusicView();
       } else {
         if (_scoreManager.currentScoreName == ScoreManager.UNIVERSE_SCORE ||
             _scoreManager.currentScoreName == ScoreManager.WEB_SCORE ||
@@ -622,13 +585,7 @@ class _MyHomePageState extends State<MyHomePage> with TickerProviderStateMixin {
           //   });
           // });
         }
-        if (_prevSelectedMelody != null) {
-          _selectOrDeselectMelody(_prevSelectedMelody);
-        } else if (_prevSelectedPart != null) {
-          _selectOrDeselectPart(_prevSelectedPart);
-        } else {
-          musicViewMode = MusicViewMode.section;
-        }
+        _selectOrDeselectMelody(_prevSelectedMelody);
         interactionMode = InteractionMode.edit;
         // showSections = true;
         _showMusicView();
@@ -699,7 +656,7 @@ class _MyHomePageState extends State<MyHomePage> with TickerProviderStateMixin {
 
   bool get _portraitPhoneUI => !_landscapePhoneUI && !_scalableUI;
 
-  BuildContext nativeDeviceOrientationReaderContext;
+  late BuildContext nativeDeviceOrientationReaderContext;
 
   NativeDeviceOrientation get _nativeOrientation {
     try {
@@ -841,15 +798,17 @@ class _MyHomePageState extends State<MyHomePage> with TickerProviderStateMixin {
   double get horizontalSectionListHeight =>
       showSections && !verticalSectionList ? 36 : 0;
 
-  ExportUI exportUI;
-  MessagesUI messagesUI;
-  UniverseViewUI universeViewUI;
+  late KeyboardVisibilityController keyboardVisibilityController;
+  late StreamSubscription<bool> keyboardVisbilitySubscription;
+  late ExportUI exportUI;
+  late MessagesUI messagesUI;
+  late UniverseViewUI universeViewUI;
   double get universeViewUIHeight => universeViewUI.height(context,
       keyboardHeight: _keyboardHeight, settingsHeight: _midiSettingsHeight);
-  BSMethod scrollToCurrentBeat;
-  BSMethod refreshUniverseData;
-  BSMethod bluetoothScan;
-  BSMethod duplicateCurrentScore;
+  late BSMethod scrollToCurrentBeat;
+  late BSMethod refreshUniverseData;
+  late BSMethod bluetoothScan;
+  late BSMethod duplicateCurrentScore;
 
   @override
   void initState() {
@@ -889,19 +848,21 @@ class _MyHomePageState extends State<MyHomePage> with TickerProviderStateMixin {
     _currentSection = widget.initialScore.sections[0];
     _scoreManager.doOpenScore = doOpenScore;
     if (widget.pastebinCode != null) {
-      _scoreManager.loadPastebinScoreIntoUI(widget.pastebinCode, onFail: () {
+      _scoreManager.loadPastebinScoreIntoUI(widget.pastebinCode!, onFail: () {
         messagesUI.sendMessage(message: "Failed to load URL!", isError: true);
       });
-    } else if (MyPlatform.isWeb) {
-      BeatScratchPlugin.createScore(score);
     }
-    if (MyPlatform.isMobile) {
-      KeyboardVisibility.onChange.listen((bool visible) {
+    keyboardVisibilityController = KeyboardVisibilityController();
+    keyboardVisbilitySubscription =
+        keyboardVisibilityController.onChange.listen((bool visible) {
+      // print('Keyboard visibility update. Is visible: $visible');
+      if (MyPlatform.isMobile) {
         setState(() {
           _softKeyboardVisible = visible;
         });
-      });
-    }
+      }
+    });
+
 //    BeatScratchPlugin.createScore(_score);
     BeatScratchPlugin.onSectionSelected = (sectionId) {
       setState(() {
@@ -951,10 +912,12 @@ class _MyHomePageState extends State<MyHomePage> with TickerProviderStateMixin {
             : null;
     RecordedSegmentQueue.updateRecordingMelody =
         BeatScratchPlugin.onRecordingMelodyUpdated;
-    keyboardPart = score.parts.firstWhere((part) => true, orElse: () => null);
-    colorboardPart = score.parts.firstWhere(
-        (part) => part.instrument.type == InstrumentType.harmonic,
-        orElse: () => null);
+    keyboardPart = score.parts.firstWhereOrNull(
+      (part) => true,
+    );
+    colorboardPart = score.parts.firstWhereOrNull(
+      (part) => part.instrument.type == InstrumentType.harmonic,
+    );
 
     colorboardNotesNotifier = ValueNotifier(Set());
     keyboardNotesNotifier = ValueNotifier(Set());
@@ -1089,6 +1052,8 @@ class _MyHomePageState extends State<MyHomePage> with TickerProviderStateMixin {
             ),
           )) ??
           false;
+    } else {
+      return true;
     }
   }
 
@@ -1127,10 +1092,6 @@ class _MyHomePageState extends State<MyHomePage> with TickerProviderStateMixin {
 
   @override
   Widget build(BuildContext context) {
-    if (splitMode == null) {
-      splitMode = (context.isTablet) ? SplitMode.half : SplitMode.full;
-      verticalSectionList = context.isTablet || context.isLandscapePhone;
-    }
     if (hasPrioritizedMIDIController && !hadPriotizedMIDIController) {
       showKeyboard = false;
     }
@@ -1140,9 +1101,10 @@ class _MyHomePageState extends State<MyHomePage> with TickerProviderStateMixin {
     //   verticalSectionList = context.isLandscape;
     // }
     if (context.isLandscape) {
-      SystemChrome.setEnabledSystemUIOverlays([]);
+      SystemChrome.setEnabledSystemUIMode(SystemUiMode.manual, overlays: []);
     } else {
-      SystemChrome.setEnabledSystemUIOverlays(SystemUiOverlay.values);
+      SystemChrome.setEnabledSystemUIMode(SystemUiMode.manual,
+          overlays: SystemUiOverlay.values);
     }
     if (BeatScratchPlugin.playing) {
       _tapInBeat = null;
@@ -1392,19 +1354,22 @@ class _MyHomePageState extends State<MyHomePage> with TickerProviderStateMixin {
                   padding: EdgeInsets.zero,
                   onPressed: _universeManager.redditUsername.isNotEmpty
                       ? () {
-                          bool oldValue = _universeManager
-                              .currentUniverseScoreFuture?.likes;
+                          if (scoreFuture == null) return;
+                          bool? oldValue = scoreFuture.likes;
                           setState(() {
                             if (oldValue == true) {
-                              scoreFuture?.likes = null;
-                              scoreFuture?.voteCount -= 1;
+                              scoreFuture.likes = null;
+                              if (scoreFuture.voteCount != null)
+                                scoreFuture.voteCount =
+                                    scoreFuture.voteCount! - 1;
                             } else {
-                              scoreFuture?.likes = true;
-                              scoreFuture?.voteCount +=
-                                  oldValue == null ? 1 : 2;
+                              scoreFuture.likes = true;
+                              if (scoreFuture.voteCount != null)
+                                scoreFuture.voteCount = scoreFuture.voteCount! +
+                                    (oldValue == null ? 1 : 2);
                             }
                             _universeManager.vote(
-                                scoreFuture?.fullName, scoreFuture?.likes);
+                                scoreFuture.fullName ?? '', scoreFuture.likes);
                           });
                         }
                       : null,
@@ -1428,7 +1393,7 @@ class _MyHomePageState extends State<MyHomePage> with TickerProviderStateMixin {
                 duration: animationDuration,
                 child: Align(
                     alignment: Alignment.center,
-                    child: Text(scoreFuture?.voteCount?.toString() ?? '',
+                    child: Text(scoreFuture?.voteCount.toString() ?? '',
                         textAlign: TextAlign.center,
                         style: TextStyle(
                             color: musicForegroundColor,
@@ -1440,18 +1405,20 @@ class _MyHomePageState extends State<MyHomePage> with TickerProviderStateMixin {
                   padding: EdgeInsets.zero,
                   onPressed: _universeManager.redditUsername.isNotEmpty
                       ? () {
-                          bool oldValue = _universeManager
-                              .currentUniverseScoreFuture?.likes;
+                          if (scoreFuture == null) return;
+                          bool? oldValue = scoreFuture.likes;
                           setState(() {
                             if (oldValue == false) {
-                              scoreFuture?.likes = null;
-                              scoreFuture.voteCount += 1;
+                              scoreFuture.likes = null;
+                              scoreFuture.voteCount =
+                                  scoreFuture.voteCount! + 1;
                             } else {
-                              scoreFuture?.likes = false;
-                              scoreFuture.voteCount -= oldValue == null ? 1 : 2;
+                              scoreFuture.likes = false;
+                              scoreFuture.voteCount = scoreFuture.voteCount! -
+                                  (oldValue == null ? 1 : 2);
                             }
                             _universeManager.vote(
-                                scoreFuture?.fullName, scoreFuture?.likes);
+                                scoreFuture.fullName ?? '', scoreFuture.likes);
                           });
                         }
                       : null,
@@ -1511,13 +1478,16 @@ class _MyHomePageState extends State<MyHomePage> with TickerProviderStateMixin {
                 child: MyFlatButton(
                   padding: EdgeInsets.zero,
                   onPressed: () {
+                    if (scoreFuture == null) return;
+
                     if (_appSettings.enableApollo) {
                       launchURL(
-                          scoreFuture.commentUrl
-                              .replaceAll("https://", "apollo://"),
+                          scoreFuture?.commentUrl
+                                  ?.replaceAll("https://", "apollo://") ??
+                              '',
                           forceSafariVC: false);
-                    } else {
-                      launchURL(scoreFuture.commentUrl, forceSafariVC: false);
+                    } else if (scoreFuture.commentUrl != null) {
+                      launchURL(scoreFuture.commentUrl!, forceSafariVC: false);
                     }
                   },
                   child: Align(
@@ -1765,16 +1735,13 @@ class _MyHomePageState extends State<MyHomePage> with TickerProviderStateMixin {
     bool bottom = false,
   }) {
     bool playing = BeatScratchPlugin.playing;
-    int tapInBeat = _tapInBeat;
+    final int? tapInBeat = _tapInBeat;
     bool isDisplayed = (!vertical && !bottom && _tapInBarHeight != 0) ||
         (vertical && _landscapeTapInBarWidth != 0) ||
         (bottom && _bottomTapInBarHeight != 0);
-    final double tapInFirstSize =
-        !playing && (vertical || tapInBeat == null) ? 42 : 0;
+    final double tapInFirstSize = !playing && (vertical) ? 42 : 0;
     final double tapInSecondSize =
-        !BeatScratchPlugin.playing && (_tapInBeat == null || _tapInBeat <= -2)
-            ? 42
-            : 0;
+        !playing && (tapInBeat != null) && (tapInBeat <= -2) ? 42 : 0;
     final audioButtonColor =
         BeatScratchPlugin.metronomeEnabled ? sectionColor : Colors.grey;
     Widget tapInBarInstructions({bool withText = true, bool withIcon = true}) =>
@@ -2245,12 +2212,10 @@ class _MyHomePageState extends State<MyHomePage> with TickerProviderStateMixin {
                   showViewOptions;
               showMidiConfiguration = true;
               showViewOptions = true;
-              if (keyboardPart != null && showKeyboard) {
+              if (showKeyboard) {
                 _showKeyboardConfiguration = true;
               }
-              if (_enableColorboard &&
-                  colorboardPart != null &&
-                  showColorboard) {
+              if (_enableColorboard && showColorboard) {
                 _showColorboardConfiguration = true;
               }
             });
@@ -2263,14 +2228,8 @@ class _MyHomePageState extends State<MyHomePage> with TickerProviderStateMixin {
           },
           saveCurrentScore: saveCurrentScore,
           pasteScore: () async {
-            ClipboardData data = await Clipboard.getData(Clipboard.kTextPlain);
-            if (data == null) {
-              setState(() {
-                pasteFailed = true;
-              });
-              return;
-            }
-            String scoreUrl = data.text;
+            ClipboardData? data = await Clipboard.getData(Clipboard.kTextPlain);
+            String scoreUrl = data?.text ?? '';
             _scoreManager.loadFromScoreUrl(scoreUrl,
                 currentScoreToSave: this.score, onFail: () {
               setState(() {
@@ -2333,11 +2292,9 @@ class _MyHomePageState extends State<MyHomePage> with TickerProviderStateMixin {
           _selectOrDeselectPart(object);
         }
       } else if (object is Section) {
-        if (selectedMelody != null) {
-          _selectOrDeselectMelody(selectedMelody);
-        }
+        _selectOrDeselectMelody(selectedMelody);
         if (selectedPart != null) {
-          _selectOrDeselectPart(selectedPart);
+          _selectOrDeselectPart(selectedPart!);
         }
         musicViewMode = MusicViewMode.section;
         if (!interactionMode.isEdit) {
@@ -2407,9 +2364,8 @@ class _MyHomePageState extends State<MyHomePage> with TickerProviderStateMixin {
         });
       },
       showTempoConfiguration: _showTapInBar,
-      visible: (_secondToolbarHeight != null && _secondToolbarHeight != 0) ||
-          (_landscapePhoneSecondToolbarWidth != null &&
-              _landscapePhoneSecondToolbarWidth != 0) ||
+      visible: (_secondToolbarHeight != 0) ||
+          (_landscapePhoneSecondToolbarWidth != 0) ||
           _scalableUI,
       tempoLongPress: _landscapePhoneUI
           ? () {
@@ -2453,7 +2409,7 @@ class _MyHomePageState extends State<MyHomePage> with TickerProviderStateMixin {
     return result;
   }
 
-  saveCurrentScore({Duration delay}) {
+  saveCurrentScore({Duration? delay}) {
     final score = this.score;
     final scoreFile = _scoreManager.currentScoreFile;
     print("Saving score ${score.name} to ${scoreFile.path.split('/').last}...");
@@ -2474,9 +2430,8 @@ class _MyHomePageState extends State<MyHomePage> with TickerProviderStateMixin {
 
     if (delay == null) {
       Future.microtask(doSave);
-    } else {
+    } else
       Future.delayed(delay, doSave);
-    }
   }
 
   double beatScratchToolbarHeight(BuildContext context) =>
@@ -2698,8 +2653,8 @@ class _MyHomePageState extends State<MyHomePage> with TickerProviderStateMixin {
         setPartVolume: _setPartVolume,
         setColorboardPart: _setColorboardPart,
         setKeyboardPart: _setKeyboardPart,
-        colorboardPart: colorboardPart,
-        keyboardPart: keyboardPart,
+        colorboardPart: colorboardPart ?? Part(),
+        keyboardPart: keyboardPart ?? Part(),
         editingMelody: recordingMelody,
         hideMelodyView: _hideMusicView,
         availableWidth: availableWidth,
@@ -2836,8 +2791,7 @@ class _MyHomePageState extends State<MyHomePage> with TickerProviderStateMixin {
               }
             : null,
         cloneCurrentSection: () {
-          if (currentSection.name == null ||
-              currentSection.name.trim().isEmpty) {
+          if (currentSection.name.trim().isEmpty) {
             String prefix = "Section";
             while (score.sections.any((s) => s.name.startsWith("$prefix "))) {
               prefix = "$prefix'";
@@ -2849,9 +2803,9 @@ class _MyHomePageState extends State<MyHomePage> with TickerProviderStateMixin {
           final match = RegExp(
             r"^(.*?)(\d*)\s*$",
           ).allMatches(section.name).first;
-          String prefix = match.group(1);
+          String prefix = match.group(1)!;
           prefix = prefix.trim();
-          int number = int.tryParse(match.group(2)) ?? 1;
+          int number = int.tryParse(match.group(2)!) ?? 1;
           while (score.sections.any((s) => s.name == section.name)) {
             section.name = "$prefix ${++number}";
           }
@@ -3092,6 +3046,11 @@ class _MyHomePageState extends State<MyHomePage> with TickerProviderStateMixin {
           part: keyboardPart,
           height: _keyboardHeight,
           showConfiguration: _showKeyboardConfiguration,
+          hideConfiguration: () {
+            setState(() {
+              _showKeyboardConfiguration = false;
+            });
+          },
           sectionColor: sectionColor,
           pressedNotesNotifier: keyboardNotesNotifier,
           bluetoothControllerPressedNotes: bluetoothControllerPressedNotes,
@@ -3119,6 +3078,12 @@ class _MyHomePageState extends State<MyHomePage> with TickerProviderStateMixin {
           part: colorboardPart,
           height: _colorboardHeight,
           showConfiguration: _showColorboardConfiguration,
+          hideConfiguration: () {
+            setState(() {
+              // showColorboard = false;
+              _showColorboardConfiguration = false;
+            });
+          },
           sectionColor: sectionColor,
           pressedNotesNotifier: colorboardNotesNotifier,
           distanceFromBottom:
