@@ -1,5 +1,4 @@
 import '../util/bs_methods.dart';
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
 import '../generated/protos/music.pb.dart';
@@ -17,10 +16,10 @@ class MelodyPreview extends StatefulWidget {
   final double scale;
 
   const MelodyPreview({
-    Key key,
-    this.section,
-    this.melody,
-    this.part,
+    Key? key,
+    required this.section,
+    required this.melody,
+    required this.part,
     this.width = 300,
     this.height = 100,
     this.scale = 0.15,
@@ -31,9 +30,9 @@ class MelodyPreview extends StatefulWidget {
 
   static Gradient generateVolumeDecoration(
       MelodyReference reference, Section section,
-      {@required bool isSelectedMelody,
-      @required Color bgColor,
-      @required Color sectionColor}) {
+      {required bool isSelectedMelody,
+      required Color bgColor,
+      required Color sectionColor}) {
     if (reference.isEnabled) {
       Color volumeColor = isSelectedMelody
           ? Colors.white
@@ -78,27 +77,25 @@ class MelodyPreview extends StatefulWidget {
 }
 
 class _MelodyPreviewState extends State<MelodyPreview> {
-  String lastPreviewKey;
-  Score preview;
-  BSMethod notifyUpdate;
+  late String lastPreviewKey;
+  late Score preview;
+  late BSMethod notifyUpdate;
 
   String get previewKey =>
-      "${widget.melody.id}-${widget.melody.hashCode}|${widget.part?.id ?? "null"}|" +
+      "${widget.melody.id}-${widget.melody.hashCode}|${widget.part.id}|" +
       "${widget.section.id}-${widget.section.hashCode}";
 
   @override
   initState() {
     super.initState();
-    preview = melodyPreview(widget.melody ?? Melody(), widget.part ?? Part(),
-        widget.section ?? Section());
+    preview = melodyPreview(widget.melody, widget.part, widget.section);
     notifyUpdate = BSMethod();
   }
 
   @override
   Widget build(BuildContext context) {
     if (lastPreviewKey != previewKey) {
-      preview = melodyPreview(widget.melody ?? Melody(), widget.part ?? Part(),
-          widget.section ?? Section());
+      preview = melodyPreview(widget.melody, widget.part, widget.section);
       lastPreviewKey = previewKey;
       notifyUpdate();
     }
