@@ -5,7 +5,6 @@ import 'package:flutter/material.dart';
 import '../generated/protos/music.pb.dart';
 import 'score_preview.dart';
 import '../util/dummydata.dart';
-import '../util/music_theory.dart';
 import '../colors.dart';
 
 class SectionPreview extends StatefulWidget {
@@ -16,9 +15,9 @@ class SectionPreview extends StatefulWidget {
   final double scale;
 
   const SectionPreview({
-    Key key,
-    this.section,
-    this.score,
+    Key? key,
+    required this.section,
+    required this.score,
     this.width = 300,
     this.height = 100,
     this.scale = 0.15,
@@ -29,9 +28,9 @@ class SectionPreview extends StatefulWidget {
 }
 
 class _SectionPreviewState extends State<SectionPreview> {
-  String lastPreviewKey;
-  Score preview;
-  BSMethod notifyUpdate;
+  String? lastPreviewKey;
+  late Score preview;
+  late BSMethod notifyUpdate;
 
   String get previewKey => "${widget.section.id}-${widget.section.hashCode}";
 
