@@ -13,7 +13,7 @@ import '../widget/my_buttons.dart';
 class MelodyToolbar extends StatefulWidget {
   final MusicViewMode musicViewMode;
   final bool editingMelody;
-  final Melody melody;
+  final Melody? melody;
   final Section currentSection;
   final Color sectionColor;
   final Function(MelodyReference) toggleMelodyReference;
@@ -23,18 +23,18 @@ class MelodyToolbar extends StatefulWidget {
   final Function(Melody) deleteMelody;
 
   const MelodyToolbar(
-      {Key key,
-      this.melody,
-      this.currentSection,
-      this.toggleMelodyReference,
-      this.setReferenceVolume,
-      this.editingMelody,
-      this.sectionColor,
-      this.toggleRecording,
-      this.backToPart,
-      this.setMelodyName,
-      this.musicViewMode,
-      this.deleteMelody})
+      {Key? key,
+      required this.melody,
+      required this.currentSection,
+      required this.toggleMelodyReference,
+      required this.setReferenceVolume,
+      required this.editingMelody,
+      required this.sectionColor,
+      required this.toggleRecording,
+      required this.backToPart,
+      required this.setMelodyName,
+      required this.musicViewMode,
+      required this.deleteMelody})
       : super(key: key);
 
   @override
@@ -42,17 +42,17 @@ class MelodyToolbar extends StatefulWidget {
 }
 
 class MelodyToolbarState extends State<MelodyToolbar> {
-  bool _showVolume;
-  TextEditingController nameController;
+  late bool _showVolume;
+  late TextEditingController nameController;
 
-  MelodyReference get melodyReference =>
-      widget.currentSection.referenceTo(widget.melody);
+  MelodyReference? get melodyReference => widget.melody != null
+      ? widget.currentSection.referenceTo(widget.melody!)
+      : null;
   bool get melodySelected => widget.melody != null;
   bool get melodyEnabled =>
       melodySelected && (melodyReference?.isEnabled ?? false);
-  Melody confirmingDeleteFor;
-  bool get isConfirmingDelete =>
-      confirmingDeleteFor != null && confirmingDeleteFor == widget.melody;
+  Melody? confirmingDeleteFor;
+  bool get isConfirmingDelete => confirmingDeleteFor == widget.melody;
   bool get showVolume =>
       (melodyReference?.isEnabled == true) &&
       (/*context.isTablet ||*/ _showVolume);
@@ -76,11 +76,9 @@ class MelodyToolbarState extends State<MelodyToolbar> {
     if (context.isTabletOrLandscapey) {
       width = width / 2;
     }
-    _showVolume &= melodyReference != null &&
-        (melodyReference?.isEnabled == true) &&
-        !isConfirmingDelete;
+    _showVolume &= (melodyReference?.isEnabled == true) && !isConfirmingDelete;
 
-    if (confirmingDeleteFor != null && confirmingDeleteFor != widget.melody) {
+    if (confirmingDeleteFor != widget.melody) {
       confirmingDeleteFor = null;
     }
     nameController.value =
@@ -98,9 +96,11 @@ class MelodyToolbarState extends State<MelodyToolbar> {
                       textCapitalization: TextCapitalization.words,
                       onChanged: (melodySelected)
                           ? (value) {
-                              widget.melody.name = value;
-                              widget.setMelodyName(
-                                  widget.melody, widget.melody.name);
+                              final melody = widget.melody;
+                              if (melody == null) return;
+
+                              melody.name = value;
+                              widget.setMelodyName(melody, melody.name);
 //                        BeatScratchPlugin.updateMelody(widget.melody);
                             }
                           : null,
@@ -109,7 +109,7 @@ class MelodyToolbarState extends State<MelodyToolbar> {
 //                      },
                       decoration: InputDecoration(
                         border: InputBorder.none,
-                        hintText: (melodySelected) ? widget.melody.idName : "",
+                        hintText: (melodySelected) ? widget.melody?.idName : "",
                       ))
                   : Text(""))),
       // AnimatedContainer(
@@ -166,7 +166,7 @@ class MelodyToolbarState extends State<MelodyToolbar> {
               onChanged: (melodyReference?.isEnabled != true)
                   ? null
                   : (value) {
-                      widget.setReferenceVolume(melodyReference, value);
+                      widget.setReferenceVolume(melodyReference!, value);
                     }),
         ),
       ),
@@ -178,7 +178,7 @@ class MelodyToolbarState extends State<MelodyToolbar> {
           child: MyRaisedButton(
               onPressed: melodySelected
                   ? () {
-                      widget.toggleMelodyReference(melodyReference);
+                      widget.toggleMelodyReference(melodyReference!);
                     }
                   : null,
               onLongPress: (melodyReference?.isEnabled == true)
@@ -214,7 +214,7 @@ class MelodyToolbarState extends State<MelodyToolbar> {
           child: MyRaisedButton(
               onPressed: () {
                 setState(() {
-                  widget.deleteMelody(confirmingDeleteFor);
+                  widget.deleteMelody(confirmingDeleteFor!);
                   confirmingDeleteFor = null;
                 });
               },
@@ -245,12 +245,14 @@ class MelodyToolbarState extends State<MelodyToolbar> {
           child: MyRaisedButton(
               onPressed: () {
                 setState(() {
-                  if (widget.melody.name?.isEmpty != false &&
-                      (widget.melody.midiData.data.isEmpty ||
-                          widget.melody.midiData.data.values
-                              .where((mc) => mc.data.length != 0)
-                              .isEmpty)) {
-                    widget.deleteMelody(widget.melody);
+                  if (widget.melody?.name.isEmpty != false &&
+                      (widget.melody?.midiData.data.isEmpty ??
+                          false ||
+                              (widget.melody?.midiData.data.values
+                                      .where((mc) => mc.data.length != 0)
+                                      .isEmpty ??
+                                  false))) {
+                    widget.deleteMelody(widget.melody!);
                   } else {
                     confirmingDeleteFor = widget.melody;
                   }
@@ -269,7 +271,7 @@ class MelodyToolbarState extends State<MelodyToolbar> {
 
 class PartToolbar extends StatefulWidget {
   final Color sectionColor;
-  final Part part;
+  final Part? part;
   final Function(Part) setKeyboardPart;
   final Function(Part) setColorboardPart;
   final Part colorboardPart;
@@ -281,19 +283,19 @@ class PartToolbar extends StatefulWidget {
   final bool enableColorboard;
 
   const PartToolbar(
-      {Key key,
-      this.part,
-      this.setKeyboardPart,
-      this.setColorboardPart,
-      this.colorboardPart,
-      this.keyboardPart,
-      this.deletePart,
-      this.configuringPart,
-      this.browsingPartMelodies,
-      this.toggleConfiguringPart,
-      this.sectionColor,
-      this.enableColorboard,
-      this.toggleBrowsingPartMelodies})
+      {Key? key,
+      required this.part,
+      required this.setKeyboardPart,
+      required this.setColorboardPart,
+      required this.colorboardPart,
+      required this.keyboardPart,
+      required this.deletePart,
+      required this.configuringPart,
+      required this.browsingPartMelodies,
+      required this.toggleConfiguringPart,
+      required this.sectionColor,
+      required this.enableColorboard,
+      required this.toggleBrowsingPartMelodies})
       : super(key: key);
 
   @override
@@ -301,17 +303,16 @@ class PartToolbar extends StatefulWidget {
 }
 
 class PartToolbarState extends State<PartToolbar> {
-  Part confirmingDeleteFor;
+  Part? confirmingDeleteFor;
 
-  bool get isConfirmingDelete =>
-      confirmingDeleteFor != null && confirmingDeleteFor == widget.part;
+  bool get isConfirmingDelete => confirmingDeleteFor == widget.part;
 
   Widget build(BuildContext context) {
     var width = MediaQuery.of(context).size.width;
     if (context.isTabletOrLandscapey) {
       width = width / 2;
     }
-    if (confirmingDeleteFor != null && confirmingDeleteFor != widget.part) {
+    if (confirmingDeleteFor != widget.part) {
       confirmingDeleteFor = null;
     }
     return Container(
@@ -328,7 +329,7 @@ class PartToolbarState extends State<PartToolbar> {
                 color: widget.configuringPart ? Colors.white : null,
                 child: AnimatedOpacity(
                     duration: animationDuration,
-                    opacity: widget.part == null || isConfirmingDelete ? 0 : 1,
+                    opacity: isConfirmingDelete ? 0 : 1,
                     child: Icon(Icons.settings,
                         color: widget.configuringPart
                             ? Colors.black
@@ -336,7 +337,7 @@ class PartToolbarState extends State<PartToolbar> {
         Expanded(
             child: Padding(
                 padding: EdgeInsets.only(left: 5),
-                child: Text((widget.part != null) ? widget.part.midiName : "",
+                child: Text((widget.part != null) ? widget.part!.midiName : "",
                     maxLines: 1,
                     overflow: TextOverflow.ellipsis,
                     style: TextStyle(
@@ -377,28 +378,23 @@ class PartToolbarState extends State<PartToolbar> {
             height: 36,
             padding: EdgeInsets.only(right: 5),
             child: MyRaisedButton(
-                onPressed: (widget.part != null &&
-                        widget.part.instrument.type != InstrumentType.drum)
+                onPressed: (widget.part!.instrument.type != InstrumentType.drum)
                     ? () {
-                        widget.setColorboardPart(widget.part);
+                        widget.setColorboardPart(widget.part!);
                       }
                     : null,
                 padding: EdgeInsets.zero,
                 child: AnimatedOpacity(
                     duration: animationDuration,
-                    opacity: widget.part == null ||
-                            isConfirmingDelete ||
-                            !widget.enableColorboard
-                        ? 0
-                        : 1,
+                    opacity:
+                        isConfirmingDelete || !widget.enableColorboard ? 0 : 1,
                     child: Stack(children: [
                       Align(
                           alignment: Alignment.bottomRight,
                           child: AnimatedOpacity(
                               duration: animationDuration,
-                              opacity: (widget.part != null &&
-                                      widget.part.instrument.type !=
-                                          InstrumentType.drum)
+                              opacity: (widget.part!.instrument.type !=
+                                      InstrumentType.drum)
                                   ? 1
                                   : 0.25,
                               child: Padding(
@@ -431,7 +427,7 @@ class PartToolbarState extends State<PartToolbar> {
               padding: EdgeInsets.all(0),
               child: AnimatedOpacity(
                   duration: animationDuration,
-                  opacity: (widget.part != null && !isConfirmingDelete) ? 0 : 0,
+                  opacity: (!isConfirmingDelete) ? 0 : 0,
                   child: Stack(children: [
                     Align(
                         alignment: Alignment.topLeft,
@@ -463,7 +459,7 @@ class PartToolbarState extends State<PartToolbar> {
             child: MyRaisedButton(
                 onPressed: () {
                   setState(() {
-                    widget.deletePart(widget.part);
+                    widget.deletePart(widget.part!);
                     confirmingDeleteFor = null;
                   });
                 },
@@ -517,23 +513,23 @@ class SectionToolbar extends StatefulWidget {
   final MusicViewMode musicViewMode;
   final Function(Section, String) setSectionName;
   final Function(Section) deleteSection;
-  final Function addPart;
+  final Function? addPart;
   final Function cloneCurrentSection;
   final bool editingSection;
   final Function(bool) setEditingSection;
 
   const SectionToolbar(
-      {Key key,
-      this.currentSection,
-      this.sectionColor,
-      this.musicViewMode,
-      this.setSectionName,
-      this.deleteSection,
-      this.canDeleteSection,
-      this.editingSection,
-      this.addPart,
-      this.cloneCurrentSection,
-      this.setEditingSection})
+      {Key? key,
+      required this.currentSection,
+      required this.sectionColor,
+      required this.musicViewMode,
+      required this.setSectionName,
+      required this.deleteSection,
+      required this.canDeleteSection,
+      required this.editingSection,
+      required this.addPart,
+      required this.cloneCurrentSection,
+      required this.setEditingSection})
       : super(key: key);
 
   @override
@@ -541,11 +537,9 @@ class SectionToolbar extends StatefulWidget {
 }
 
 class SectionToolbarState extends State<SectionToolbar> {
-  Section confirmingDeleteFor;
+  Section? confirmingDeleteFor;
 
-  bool get isConfirmingDelete =>
-      confirmingDeleteFor != null &&
-      confirmingDeleteFor == widget.currentSection;
+  bool get isConfirmingDelete => confirmingDeleteFor == widget.currentSection;
 
   TextEditingController nameController = TextEditingController();
   @override
@@ -560,8 +554,7 @@ class SectionToolbarState extends State<SectionToolbar> {
     if (context.isTabletOrLandscapey) {
       width = width / 2;
     }
-    if (confirmingDeleteFor != null &&
-        confirmingDeleteFor != widget.currentSection) {
+    if (confirmingDeleteFor != widget.currentSection) {
       confirmingDeleteFor = null;
     }
     nameController.value =
@@ -619,17 +612,16 @@ class SectionToolbarState extends State<SectionToolbar> {
           )),
       AnimatedContainer(
           duration: animationDuration,
-          width: isConfirmingDelete || widget.addPart == null ? 0 : 69,
+          width: isConfirmingDelete ? 0 : 69,
           height: 36,
           padding: EdgeInsets.only(right: 5),
           child: MyRaisedButton(
-              onPressed: widget.addPart,
+              onPressed: widget.addPart as VoidCallback?,
               padding: EdgeInsets.zero,
               child: AnimatedOpacity(
                   duration: animationDuration,
                   opacity: widget.musicViewMode != MusicViewMode.section ||
-                          isConfirmingDelete ||
-                          widget.addPart == null
+                          isConfirmingDelete
                       ? 0
                       : 1,
                   child: Row(children: [
@@ -648,7 +640,7 @@ class SectionToolbarState extends State<SectionToolbar> {
           height: 36,
           padding: EdgeInsets.only(right: 5),
           child: MyRaisedButton(
-              onPressed: widget.cloneCurrentSection,
+              onPressed: widget.cloneCurrentSection as VoidCallback,
               padding: EdgeInsets.zero,
               child: AnimatedOpacity(
                   duration: animationDuration,
@@ -676,7 +668,7 @@ class SectionToolbarState extends State<SectionToolbar> {
           child: MyRaisedButton(
               onPressed: () {
                 setState(() {
-                  widget.deleteSection(confirmingDeleteFor);
+                  widget.deleteSection(confirmingDeleteFor!);
                   confirmingDeleteFor = null;
                 });
               },
