@@ -1,7 +1,6 @@
 import 'dart:convert';
 import 'dart:io';
 import 'dart:math';
-import 'dart:typed_data';
 
 import 'package:flutter/foundation.dart';
 import 'package:http/http.dart' as http;
@@ -27,30 +26,28 @@ class ScoreManager {
   static const String FROM_WEB = " (from Link)";
   static const String UNIVERSE_SCORE = "Universe Score";
   static const String FROM_UNIVERSE = " (from Universe)";
-  Function(Score) doOpenScore;
-  Directory scoresDirectory;
-  SharedPreferences _prefs;
+  late Function(Score) doOpenScore;
+  late Directory scoresDirectory;
+  late SharedPreferences _prefs;
 
   String get currentScoreName =>
-      _prefs?.getString('currentScoreName') ?? UNIVERSE_SCORE;
+      _prefs.getString('currentScoreName') ?? UNIVERSE_SCORE;
 
   set currentScoreName(String value) =>
-      _prefs?.setString("currentScoreName", value);
+      _prefs.setString("currentScoreName", value);
 
   File get currentScoreFile => File(
       "${scoresDirectory.path}/${Uri.encodeComponent(currentScoreName).replaceAll("%20", " ")}.beatscratch");
 
   List<FileSystemEntity> get scoreFiles {
-    if (scoresDirectory != null) {
-      List<FileSystemEntity> result = scoresDirectory
-          ?.listSync()
-          .where((f) => f.path.endsWith(".beatscratch"))
-          .toList();
-      result.sort(
-          (a, b) => b.statSync().modified.compareTo(a.statSync().modified));
-      return result;
-    }
-    return [];
+    List<FileSystemEntity> result = scoresDirectory
+        .listSync()
+        .where((f) => f.path.endsWith(".beatscratch"))
+        .toList();
+    result
+        .sort((a, b) => b.statSync().modified.compareTo(a.statSync().modified));
+    return result;
+    // return [];
   }
 
   ScoreManager() {
@@ -76,7 +73,7 @@ class ScoreManager {
     }
   }
 
-  createScore(String name, {Score score}) {
+  createScore(String name, {Score? score}) {
     score = score ?? defaultScore();
     currentScoreName = name;
     saveCurrentScore(score);
@@ -109,15 +106,15 @@ class ScoreManager {
     } catch (e) {
       score = defaultScore();
     }
-    doOpenScore?.call(score);
+    doOpenScore.call(score);
   }
 
   loadFromScoreUrl(String scoreUrl,
       {String newScoreDefaultFilename = PASTED_SCORE,
       String newScoreNameSuffix = FROM_CLIPBOARD,
-      Score currentScoreToSave,
-      VoidCallback onFail,
-      Function(String) onSuccess}) {
+      required Score currentScoreToSave,
+      VoidCallback? onFail,
+      Function(String)? onSuccess}) {
     print("ScoreURL=$scoreUrl");
     scoreUrl = scoreUrl.replaceFirst(new RegExp(r'http.*#score='), '');
     scoreUrl = scoreUrl.replaceFirst(new RegExp(r'http.*#/score/'), '');
@@ -126,8 +123,8 @@ class ScoreManager {
       if (scoreUrl.length < 10) {
         throw Exception("nope");
       }
-      Score score = scoreFromUrlHashValue(scoreUrl);
-      if (score == null || score.sections.isEmpty) {
+      Score score = scoreFromUrlHashValue(scoreUrl)!;
+      if (score.sections.isEmpty) {
         throw Exception("nope");
       }
       String scoreName = score.name ?? "";
@@ -137,9 +134,7 @@ class ScoreManager {
       } else {
         suggestedScoreName += newScoreNameSuffix;
       }
-      if (currentScoreToSave != null) {
-        saveCurrentScore(currentScoreToSave);
-      }
+      saveCurrentScore(currentScoreToSave);
       openScoreWithFilename(
           score, newScoreDefaultFilename); // side-effect: updates this.score
       _lastSuggestedScoreName = suggestedScoreName;
@@ -154,8 +149,8 @@ class ScoreManager {
     }
   }
 
-  static String _lastSuggestedScoreName;
-  static String get lastSuggestedScoreName {
+  static String? _lastSuggestedScoreName;
+  static String? get lastSuggestedScoreName {
     final value = _lastSuggestedScoreName;
     _lastSuggestedScoreName = null;
     return value;
@@ -166,7 +161,7 @@ class ScoreManager {
   }
 
   Future<Score> loadPastebinScore(String codeOrUrl,
-      {String titleOverride}) async {
+      {String? titleOverride}) async {
     final code = codeOrUrl.replaceFirst(new RegExp(r'http.*#/s/'), '');
 
     http.Response response = await http.get(
@@ -180,22 +175,17 @@ class ScoreManager {
     longUrl = longUrl.replaceFirst(new RegExp(r'http.*#score='), '');
     longUrl = longUrl.replaceFirst(new RegExp(r'http.*#/score/'), '');
 
-    Score score = scoreFromUrlHashValue(longUrl);
-    if (titleOverride != null) {
-      score.name = titleOverride;
-    }
+    Score score = scoreFromUrlHashValue(longUrl)!;
+    if (titleOverride != null) score.name = titleOverride;
     return score;
   }
 
   loadPastebinScoreIntoUI(String pastebinCode,
       {String newScoreDefaultFilename = PASTED_SCORE,
       String newScoreNameSuffix = FROM_CLIPBOARD,
-      Score currentScoreToSave,
-      VoidCallback onFail,
-      Function(String) onSuccess}) async {
-    if (pastebinCode == null) {
-      return;
-    }
+      Score? currentScoreToSave,
+      VoidCallback? onFail,
+      Function(String)? onSuccess}) async {
     try {
       Score score = await loadPastebinScore(pastebinCode);
       String scoreName = score.name ?? "";
@@ -234,7 +224,7 @@ class ScoreManager {
 
 extension ScoreName on FileSystemEntity {
   String get scoreName {
-    String fileName = path.split("/")?.last ?? ".beatscratch";
+    String fileName = path.split("/").last ?? ".beatscratch";
     fileName = fileName.substring(0, max(0, fileName.length - 12));
     String scoreName = Uri.decodeComponent(fileName.replaceAll(" ", "%20"));
     return scoreName;
