@@ -1,30 +1,23 @@
 import 'dart:io';
 import 'dart:math';
-import 'dart:ui';
 import 'package:beatscratch_flutter_redux/settings/app_settings.dart';
-import 'package:beatscratch_flutter_redux/storage/url_conversions.dart';
 import 'package:beatscratch_flutter_redux/universe_view/universe_icon.dart';
 import 'package:flutter/services.dart';
-import 'package:http/http.dart' as http;
 
-import 'package:beatscratch_flutter_redux/music_view/music_system_painter.dart';
 import 'package:beatscratch_flutter_redux/storage/score_picker_preview.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
-import 'dart:convert';
 
 import '../beatscratch_plugin.dart';
 import '../colors.dart';
 import '../ui_models.dart';
 import '../util/dummydata.dart';
 import '../util/bs_methods.dart';
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:animated_list_plus/animated_list_plus.dart';
 import 'package:animated_list_plus/transitions.dart';
 
 import '../generated/protos/music.pb.dart';
 import '../generated/protos/protobeats_plugin.pb.dart';
-import '../music_preview/score_preview.dart';
 import '../util/music_utils.dart';
 import '../widget/my_buttons.dart';
 import 'score_manager.dart';
@@ -53,7 +46,7 @@ extension ShowScoreNameEntry on ScorePickerMode {
 class ScorePicker extends StatefulWidget {
   final Axis scrollDirection;
   final Color sectionColor;
-  final Function(VoidCallback) setState;
+  final Function(VoidCallback)? setState;
   final VoidCallback close;
   final VoidCallback switchToUniverse;
   final ScorePickerMode mode;
@@ -67,22 +60,22 @@ class ScorePicker extends StatefulWidget {
   final BSMethod refreshUniverseData;
 
   const ScorePicker(
-      {Key key,
+      {Key? key,
       this.scrollDirection = Axis.horizontal,
-      this.sectionColor,
+      required this.sectionColor,
       this.setState,
-      this.close,
-      this.mode,
-      this.openedScore,
-      this.scoreManager,
-      this.universeManager,
-      this.appSettings,
-      this.requestKeyboardFocused,
-      this.requestMode,
-      this.width,
-      this.height,
-      this.refreshUniverseData,
-      this.switchToUniverse})
+      required this.close,
+      required this.mode,
+      required this.openedScore,
+      required this.scoreManager,
+      required this.universeManager,
+      required this.appSettings,
+      required this.requestKeyboardFocused,
+      required this.requestMode,
+      required this.width,
+      required this.height,
+      required this.refreshUniverseData,
+      required this.switchToUniverse})
       : super(key: key);
 
   @override
@@ -98,9 +91,9 @@ class ScorePickerState extends State<ScorePicker> {
   FocusNode nameFocus = FocusNode();
 
   ScoreManager get scoreManager => widget.scoreManager;
-  ScorePickerMode previousMode;
-  String deletingScoreName;
-  String overwritingScoreName;
+  ScorePickerMode? previousMode;
+  String? deletingScoreName;
+  String? overwritingScoreName;
   BSMethod universeAnimation = BSMethod();
 
   bool get wasShowingScoreNameEntry =>
@@ -187,7 +180,7 @@ class ScorePickerState extends State<ScorePicker> {
       text: '',
       style: detailTextStyle,
     ); //ing extraDetailText = "";
-    IconData icon;
+    late IconData icon;
     switch (widget.mode) {
       case ScorePickerMode.open:
         operationText = "Open";
@@ -224,7 +217,7 @@ class ScorePickerState extends State<ScorePicker> {
     }
     bool detailsTextInColumn = true; //MediaQuery.of(context).size.width < 500;
     bool showDetailsText = detailsTextInColumn &&
-        extraDetailText.text.isNotEmpty &&
+        (extraDetailText.text?.isNotEmpty ?? false) &&
         !hideDetailsText;
 
     return Column(
@@ -263,7 +256,7 @@ class ScorePickerState extends State<ScorePicker> {
                                     animateIcon: universeAnimation,
                                   )),
                             )),
-                        if (icon != null && operationText.isNotEmpty)
+                        if (operationText.isNotEmpty)
                           MyFlatButton(
                               padding: EdgeInsets.zero,
                               lightHighlight: true,
@@ -394,7 +387,7 @@ class ScorePickerState extends State<ScorePicker> {
                 ))),
         AnimatedContainer(
             height: showDetailsText
-                ? 0 == extraDetailText.children?.length ?? 0
+                ? 0 == (extraDetailText.children?.length ?? 0)
                     ? 32
                     : 48
                 : 0,
@@ -626,18 +619,15 @@ class ScorePickerState extends State<ScorePicker> {
       return widget.universeManager.cachedUniverseData;
     } else {
       return scoreManager.scoreFiles.map((scoreFile) {
-        Future<Score> loadScore() async {
-          if (scoreFile == null) {
-            return Future.value(defaultScore());
-          }
-          try {
-            final data = await File(scoreFile?.path).readAsBytes();
+        // Future<Score> loadScore() async {
+        //   try {
+        //     final data = await File(scoreFile.path).readAsBytes();
 
-            return Score.fromBuffer(data);
-          } catch (e) {
-            return Future.value(defaultScore());
-          }
-        }
+        //     return Score.fromBuffer(data);
+        //   } catch (e) {
+        //     return Future.value(defaultScore());
+        //   }
+        // }
 
         return ScoreFuture(filePath: scoreFile.path);
       }).toList();
@@ -652,15 +642,18 @@ class ScorePickerState extends State<ScorePicker> {
       spawnIsolate: false,
       controller: _scrollController,
       items: scores,
-      areItemsTheSame: (a, b) => a?.identity == b?.identity,
+      areItemsTheSame: (a, b) => a.identity == b.identity,
       // Called, as needed, to build list item widgets.
       // List items are only built when they're scrolled into view.
       itemBuilder: (context, animation, section, index) {
-        ScoreFuture scoreFuture;
+        late final ScoreFuture? scoreFuture;
         if (index < scores.length) {
           scoreFuture = scores[index];
         }
-        File scoreFile = scoreFuture?.file;
+        final scoreFile = scoreFuture?.file;
+        if (scoreFuture == null || scoreFile == null) {
+          return SizedBox();
+        }
 
         Widget tile = ScorePickerPreview(
           sectionColor: widget.sectionColor,
@@ -673,37 +666,33 @@ class ScorePickerState extends State<ScorePicker> {
               : () {
                   switch (widget.mode) {
                     case ScorePickerMode.open:
-                      if (scoreFile != null) {
-                        widget.scoreManager.openScore(scoreFile);
-                        widget.universeManager.currentUniverseScore = "";
-                      }
+                      widget.scoreManager.openScore(scoreFile);
+                      widget.universeManager.currentUniverseScore = "";
                       break;
                     case ScorePickerMode.universe:
                       if (BeatScratchPlugin.supportsStorage) {
                         scoreManager.currentScoreName =
                             ScoreManager.UNIVERSE_SCORE;
                       }
-                      scoreFuture.loadScore(scoreManager).then((value) {
+                      scoreFuture?.loadScore(scoreManager).then((value) {
                         widget.scoreManager.doOpenScore(value);
                         widget.universeManager.currentUniverseScore =
-                            scoreFuture.identity;
+                            scoreFuture!.identity;
                         widget.scoreManager.saveCurrentScore(value);
                       });
                       break;
                     default:
-                      String scoreName = scoreFile?.scoreName;
-                      if (scoreName != null) {
-                        nameController.clear();
-                        setState(() {
-                          nameController.value =
-                              nameController.value.copyWith(text: scoreName);
-                        });
-                      }
+                      String scoreName = scoreFile.scoreName;
+                      nameController.clear();
+                      setState(() {
+                        nameController.value =
+                            nameController.value.copyWith(text: scoreName);
+                      });
                   }
                 },
           scoreManager: scoreManager,
           universeManager: widget.universeManager,
-          scoreKey: (scoreFile?.lastModifiedSync() ?? DateTime(0)).hashCode,
+          scoreKey: (scoreFile.lastModifiedSync() ?? DateTime(0)).hashCode,
           scoreFuture: scoreFuture,
           deleteScore: widget.mode == ScorePickerMode.universe
               ? null
@@ -744,7 +733,7 @@ class ScorePickerState extends State<ScorePicker> {
                 bottom: widget.scrollDirection == Axis.vertical ? 10 : 0),
             child: tile);
         return SizeFadeTransition(
-            key: ValueKey("ScorePickerTile-${scoreFuture?.identity}"),
+            key: ValueKey("ScorePickerTile-${scoreFuture.identity}"),
             sizeFraction: 0.7,
             curve: Curves.easeInOut,
             axis: widget.scrollDirection,
