@@ -1,21 +1,16 @@
 import 'dart:io';
 import 'dart:math';
-import 'dart:ui';
 
-import 'package:beatscratch_flutter_redux/music_view/music_system_painter.dart';
 import 'package:beatscratch_flutter_redux/settings/app_settings.dart';
 import 'package:beatscratch_flutter_redux/util/music_notation_theory.dart';
 import 'package:beatscratch_flutter_redux/util/util.dart';
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 
-import '../beatscratch_plugin.dart';
 import '../colors.dart';
 import '../generated/protos/music.pb.dart';
 import '../music_preview/score_preview.dart';
 import '../ui_models.dart';
-import '../util/bs_methods.dart';
 import '../util/dummydata.dart';
 import '../widget/my_buttons.dart';
 import 'score_manager.dart';
@@ -23,9 +18,9 @@ import 'universe_manager.dart';
 import 'url_conversions.dart';
 
 class ScoreFuture {
-  final String filePath, scoreUrl, title, author, commentUrl, fullName;
-  int voteCount;
-  bool likes;
+  final String? filePath, scoreUrl, title, author, commentUrl, fullName;
+  int? voteCount;
+  bool? likes;
 
   ScoreFuture({
     this.filePath,
@@ -60,31 +55,28 @@ class ScoreFuture {
       };
 
   String get identity => filePath ?? "//universe-score://$scoreUrl";
-  FileSystemEntity get file {
-    if (filePath != null) {
-      try {
-        return File(filePath);
-      } catch (e) {
-        print("Error loading score from file: $e");
-      }
+  File? get file {
+    try {
+      if (filePath == null) return null;
+
+      return File(filePath!);
+    } catch (e) {
+      print("Error loading score from file: $e");
     }
     return null;
   }
 
   Future<Score> loadScore(ScoreManager scoreManager) async {
-    if (this.file != null) {
-      return loadScoreFromFile();
-    } else {
-      return loadScoreFromUniverse(scoreManager);
-    }
+    return loadScoreFromFile();
   }
 
   Future<Score> loadScoreFromFile() async {
-    if (file == null) {
-      return Future.value(defaultScore());
-    }
     try {
-      final data = await File(file?.path).readAsBytes();
+      final file = this.file;
+      if (file == null) {
+        return Future.value(defaultScore());
+      }
+      final data = await File(file.path).readAsBytes();
 
       return Score.fromBuffer(data);
     } catch (e) {
@@ -93,19 +85,19 @@ class ScoreFuture {
   }
 
   Future<Score> loadScoreFromUniverse(ScoreManager scoreManager) async {
-    String scoreUrl = this.scoreUrl;
+    String scoreUrl = this.scoreUrl!;
     scoreUrl = scoreUrl.replaceFirst(new RegExp(r'http.*#score='), '');
     scoreUrl = scoreUrl.replaceFirst(new RegExp(r'http.*#/score/'), '');
     scoreUrl = scoreUrl.replaceFirst(new RegExp(r'http.*#/s/'), '');
     try {
       final score = scoreFromUrlHashValue(scoreUrl);
       if (score == null) {
-        throw "failed to load";
+        return Future.value(defaultScore());
       }
-      return score..name = title;
+      return score..name = title!;
     } catch (e) {
       try {
-        return scoreManager.loadPastebinScore(scoreUrl, titleOverride: title);
+        return scoreManager.loadPastebinScore(scoreUrl, titleOverride: title!);
       } catch (e) {
         return Future.value(defaultScore());
       }
@@ -116,36 +108,36 @@ class ScoreFuture {
 class ScorePickerPreview extends StatefulWidget {
   final Color sectionColor;
   final ScoreFuture scoreFuture;
-  final VoidCallback deleteScore;
-  final VoidCallback overwriteScore;
+  final VoidCallback? deleteScore;
+  final VoidCallback? overwriteScore;
   final ScoreManager scoreManager;
   final UniverseManager universeManager;
   final AppSettings appSettings;
-  final VoidCallback onClickScore;
+  final VoidCallback? onClickScore;
   final int scoreKey;
-  final String deletingScoreName;
-  final String overwritingScoreName;
-  final VoidCallback cancelDelete;
-  final VoidCallback cancelOverwrite;
+  final String? deletingScoreName;
+  final String? overwritingScoreName;
+  final VoidCallback? cancelDelete;
+  final VoidCallback? cancelOverwrite;
   final double width, height;
 
   const ScorePickerPreview(
-      {Key key,
-      this.sectionColor,
-      this.scoreFuture,
-      this.deleteScore,
-      this.overwriteScore,
-      this.scoreManager,
-      this.appSettings,
+      {Key? key,
+      required this.sectionColor,
+      required this.scoreFuture,
+      required this.deleteScore,
+      required this.overwriteScore,
+      required this.scoreManager,
+      required this.appSettings,
       this.deletingScoreName,
       this.overwritingScoreName,
-      this.scoreKey,
-      this.cancelDelete,
-      this.cancelOverwrite,
-      this.universeManager,
+      required this.scoreKey,
+      required this.cancelDelete,
+      required this.cancelOverwrite,
+      required this.universeManager,
       this.onClickScore,
-      this.width,
-      this.height})
+      required this.width,
+      required this.height})
       : super(key: key);
 
   @override
@@ -153,14 +145,14 @@ class ScorePickerPreview extends StatefulWidget {
 }
 
 class _ScorePickerPreviewState extends State<ScorePickerPreview> {
-  bool _confirmingDelete;
-  bool _confirmingOverwrite;
-  int _lastScoreKey;
-  bool disposed;
+  late bool _confirmingDelete;
+  late bool _confirmingOverwrite;
+  late int _lastScoreKey;
+  late bool disposed;
 
-  Score _previewScore;
-  ScrollController scrollController;
-  BSMethod notifyUpdate;
+  late Score? _previewScore;
+  late ScrollController scrollController;
+  late BSMethod notifyUpdate;
 
   @override
   initState() {
@@ -180,9 +172,9 @@ class _ScorePickerPreviewState extends State<ScorePickerPreview> {
   }
 
   String get unloadedScoreName =>
-      widget.scoreFuture?.title ?? widget.scoreFuture?.file?.scoreName ?? "";
+      widget.scoreFuture.title ?? widget.scoreFuture.file?.scoreName ?? "";
 
-  bool get isUniverse => widget.scoreFuture?.voteCount != null;
+  bool get isUniverse => widget.scoreFuture.voteCount != null;
   @override
   Widget build(BuildContext context) {
     final scoreName = unloadedScoreName;
@@ -190,7 +182,7 @@ class _ScorePickerPreviewState extends State<ScorePickerPreview> {
         ? widget.scoreFuture.identity ==
             widget.universeManager.currentUniverseScore
         : unloadedScoreName == widget.scoreManager.currentScoreName;
-    if (widget.scoreKey != _lastScoreKey && widget.scoreFuture != null) {
+    if (widget.scoreKey != _lastScoreKey) {
       _confirmingDelete = false;
       _confirmingOverwrite = false;
       _previewScore = null;
@@ -223,17 +215,17 @@ class _ScorePickerPreviewState extends State<ScorePickerPreview> {
       backgroundColor = musicBackgroundColor;
     }
 
-    Score previewScore = _previewScore;
+    Score previewScore = _previewScore ?? defaultScore();
     // if (previewScore == null) {
     //   previewScore = defaultScore();
     // }
-    if (previewScore?.sections?.isEmpty == true) {
+    if (previewScore.sections.isEmpty == true) {
       previewScore.sections.add(defaultSection());
     }
     final actualScoreName = isUniverse
         ? unloadedScoreName
         : _previewScore != null
-            ? _previewScore.name
+            ? _previewScore!.name
             : unloadedScoreName;
 
     double previewScale =
@@ -244,8 +236,8 @@ class _ScorePickerPreviewState extends State<ScorePickerPreview> {
     return Row(children: [
       AnimatedContainer(
           duration: animationDuration,
-          width: /*widget.scoreFuture?.loadScore != null ? */ widget
-              .width /*: 0*/,
+          width: /*widget.scoreFuture?.loadScore != null ? */
+              widget.width /*: 0*/,
           height: widget.height,
           color: backgroundColor,
           padding: EdgeInsets.zero,
@@ -368,7 +360,7 @@ class _ScorePickerPreviewState extends State<ScorePickerPreview> {
                                     onPressed: () {
                                       if (!disposed) {
                                         setState(() {
-                                          widget.deleteScore();
+                                          widget.deleteScore?.call();
                                         });
                                       }
                                     },
@@ -381,7 +373,7 @@ class _ScorePickerPreviewState extends State<ScorePickerPreview> {
                                       if (!disposed) {
                                         setState(() {
                                           _confirmingDelete = false;
-                                          widget.cancelDelete();
+                                          widget.cancelDelete?.call();
                                         });
                                       }
                                     },
@@ -414,7 +406,7 @@ class _ScorePickerPreviewState extends State<ScorePickerPreview> {
                                     onPressed: () {
                                       if (!disposed) {
                                         setState(() {
-                                          widget.overwriteScore();
+                                          widget.overwriteScore?.call();
                                         });
                                       }
                                     },
@@ -427,7 +419,7 @@ class _ScorePickerPreviewState extends State<ScorePickerPreview> {
                                       if (!disposed) {
                                         setState(() {
                                           _confirmingOverwrite = false;
-                                          widget.cancelOverwrite();
+                                          widget.cancelOverwrite?.call();
                                         });
                                       }
                                     },
@@ -445,7 +437,7 @@ class _ScorePickerPreviewState extends State<ScorePickerPreview> {
                   alignment: Alignment.bottomLeft,
                   child: AnimatedOpacity(
                       duration: animationDuration,
-                      opacity: widget.scoreFuture?.author != null ? 1 : 0,
+                      opacity: widget.scoreFuture.author != null ? 1 : 0,
                       child: Container(
                           height: 36,
                           padding: EdgeInsets.all(5),
@@ -464,7 +456,7 @@ class _ScorePickerPreviewState extends State<ScorePickerPreview> {
                                       fontSize: 8))
                             ]),
                             SizedBox(width: 5),
-                            Text(widget.scoreFuture?.author ?? "",
+                            Text(widget.scoreFuture.author ?? "",
                                 style: TextStyle(
                                     color: musicForegroundColor,
                                     fontWeight: FontWeight.w700))
@@ -482,70 +474,74 @@ class _ScorePickerPreviewState extends State<ScorePickerPreview> {
               child: Column(
                 children: [
                   MyFlatButton(
-                      color: widget.scoreFuture?.likes == true
+                      color: widget.scoreFuture.likes == true
                           ? chromaticSteps[11]
                           : Colors.transparent,
                       padding: EdgeInsets.symmetric(vertical: 5),
                       onPressed:
                           widget.universeManager.redditUsername.isNotEmpty
                               ? () {
-                                  bool oldValue = widget.scoreFuture?.likes;
+                                  bool? oldValue = widget.scoreFuture.likes;
                                   setState(() {
                                     if (oldValue == true) {
-                                      widget.scoreFuture?.likes = null;
-                                      widget.scoreFuture.voteCount -= 1;
+                                      widget.scoreFuture.likes = null;
+                                      widget.scoreFuture.voteCount =
+                                          widget.scoreFuture.voteCount! - 1;
                                     } else {
-                                      widget.scoreFuture?.likes = true;
-                                      widget.scoreFuture.voteCount +=
-                                          oldValue == null ? 1 : 2;
+                                      widget.scoreFuture.likes = true;
+                                      widget.scoreFuture.voteCount =
+                                          widget.scoreFuture.voteCount! +
+                                              (oldValue == null ? 1 : 2);
                                     }
                                     widget.universeManager.vote(
-                                        widget.scoreFuture?.fullName,
-                                        widget.scoreFuture?.likes);
+                                        widget.scoreFuture.fullName!,
+                                        widget.scoreFuture.likes);
                                   });
                                 }
                               : null,
                       child: Icon(Icons.arrow_upward,
                           color: widget.universeManager.isAuthenticated
-                              ? widget.scoreFuture?.likes == true
+                              ? widget.scoreFuture.likes == true
                                   ? chromaticSteps[11].textColor()
                                   : chromaticSteps[11]
                               : musicForegroundColor.withOpacity(0.5))),
                   Row(children: [
                     Expanded(child: SizedBox()),
-                    Text(widget.scoreFuture?.voteCount?.toString() ?? '',
+                    Text(widget.scoreFuture.voteCount.toString() ?? '',
                         style: TextStyle(
                             color: musicForegroundColor,
                             fontWeight: FontWeight.w800)),
                     Expanded(child: SizedBox()),
                   ]),
                   MyFlatButton(
-                      color: widget.scoreFuture?.likes == false
+                      color: widget.scoreFuture.likes == false
                           ? chromaticSteps[10]
                           : Colors.transparent,
                       padding: EdgeInsets.symmetric(vertical: 5),
                       onPressed:
                           widget.universeManager.redditUsername.isNotEmpty
                               ? () {
-                                  bool oldValue = widget.scoreFuture?.likes;
+                                  bool? oldValue = widget.scoreFuture.likes;
                                   setState(() {
                                     if (oldValue == false) {
-                                      widget.scoreFuture?.likes = null;
-                                      widget.scoreFuture.voteCount += 1;
+                                      widget.scoreFuture.likes = null;
+                                      widget.scoreFuture.voteCount =
+                                          widget.scoreFuture.voteCount! + 1;
                                     } else {
-                                      widget.scoreFuture?.likes = false;
-                                      widget.scoreFuture.voteCount -=
-                                          oldValue == null ? 1 : 2;
+                                      widget.scoreFuture.likes = false;
+                                      widget.scoreFuture.voteCount =
+                                          widget.scoreFuture.voteCount! -
+                                              (oldValue == null ? 1 : 2);
                                     }
                                     widget.universeManager.vote(
-                                        widget.scoreFuture?.fullName,
-                                        widget.scoreFuture?.likes);
+                                        widget.scoreFuture.fullName!,
+                                        widget.scoreFuture.likes);
                                   });
                                 }
                               : null,
                       child: Icon(Icons.arrow_downward,
                           color: widget.universeManager.isAuthenticated
-                              ? widget.scoreFuture?.likes == false
+                              ? widget.scoreFuture.likes == false
                                   ? chromaticSteps[10].textColor()
                                   : chromaticSteps[10]
                               : musicForegroundColor.withOpacity(0.5))),
@@ -555,11 +551,11 @@ class _ScorePickerPreviewState extends State<ScorePickerPreview> {
                       onPressed: () {
                         if (widget.appSettings.enableApollo) {
                           launchURL(
-                              widget.scoreFuture.commentUrl
+                              widget.scoreFuture.commentUrl!
                                   .replaceAll("https://", "apollo://"),
                               forceSafariVC: false);
                         } else {
-                          launchURL(widget.scoreFuture.commentUrl,
+                          launchURL(widget.scoreFuture.commentUrl!,
                               forceSafariVC: false);
                         }
                       },
