@@ -2,64 +2,58 @@ import 'package:beatscratch_flutter_redux/messages/messages.dart';
 import 'package:beatscratch_flutter_redux/storage/score_manager.dart';
 import 'package:beatscratch_flutter_redux/storage/score_picker_preview.dart';
 import 'package:beatscratch_flutter_redux/util/util.dart';
+import 'package:collection/collection.dart';
 import 'dart:convert';
 import 'dart:io';
 
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_icons/flutter_icons.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:http/http.dart' as http;
-import 'package:path_provider/path_provider.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
-import '../beatscratch_plugin.dart';
 import '../colors.dart';
 import '../generated/protos/music.pb.dart';
-import '../util/dummydata.dart';
-import '../util/proto_utils.dart';
 import '../widget/my_platform.dart';
 import 'url_conversions.dart';
 
 class UniverseManager {
   static final REDDIT_CLIENT_ID = 'rSA9vlCRCznMCw';
   static final REDDIT_REDIRECT_URI = 'https://beatscratch.io/app';
-  Function(Score) doOpenScore;
-  Directory scoresDirectory;
-  SharedPreferences _prefs;
-  ScoreManager scoreManager;
-  MessagesUI messagesUI;
-  BSMethod refreshUniverseData;
+  late Function(Score) doOpenScore;
+  late Directory scoresDirectory;
+  late SharedPreferences _prefs;
+  late ScoreManager scoreManager;
+  late MessagesUI messagesUI;
+  late BSMethod refreshUniverseData;
 
   UniverseManager() {
     _initialize();
   }
 
-  bool get useWebViewSignIn => _prefs?.getBool('useWebViewSignIn') ?? false;
-  set useWebViewSignIn(bool v) => _prefs?.setBool("useWebViewSignIn", v);
+  bool get useWebViewSignIn => _prefs.getBool('useWebViewSignIn') ?? false;
+  set useWebViewSignIn(bool v) => _prefs.setBool("useWebViewSignIn", v);
 
   String get currentUniverseScore =>
-      _prefs?.getString('currentUniverseScore') ?? '';
+      _prefs.getString('currentUniverseScore') ?? '';
   set currentUniverseScore(String v) =>
-      _prefs?.setString("currentUniverseScore", v);
+      _prefs.setString("currentUniverseScore", v);
 
-  ScoreFuture get currentUniverseScoreFuture => currentUniverseScore == ''
+  ScoreFuture? get currentUniverseScoreFuture => currentUniverseScore == ''
       ? null
-      : cachedUniverseData.firstWhere((d) => d.identity == currentUniverseScore,
-          orElse: () => null);
+      : cachedUniverseData.firstWhereOrNull(
+          (d) => d.identity == currentUniverseScore,
+        );
 
-  String get redditRefreshToken =>
-      _prefs?.getString('redditRefreshToken') ?? "";
+  String get redditRefreshToken => _prefs.getString('redditRefreshToken') ?? "";
   set redditRefreshToken(String value) =>
-      _prefs?.setString("redditRefreshToken", value);
+      _prefs.setString("redditRefreshToken", value);
 
-  String get redditAccessToken => _prefs?.getString('redditAccessToken') ?? "";
+  String get redditAccessToken => _prefs.getString('redditAccessToken') ?? "";
   set redditAccessToken(String value) =>
-      _prefs?.setString("redditAccessToken", value);
+      _prefs.setString("redditAccessToken", value);
 
-  String get redditUsername => _prefs?.getString('redditUsername') ?? "";
-  set redditUsername(String value) =>
-      _prefs?.setString("redditUsername", value);
+  String get redditUsername => _prefs.getString('redditUsername') ?? "";
+  set redditUsername(String value) => _prefs.setString("redditUsername", value);
 
   static const String DEFAULT_UNIVERSE_DATA_STRING =
       '[{"filePath":null,"title":"Tropico-Pastoral","author":"pseudocomposer","commentUrl":"https://reddit.com/r/BeatScratch/comments/n5f1s1/tropicopastoral/","voteCount":1,"likes":true,"fullName":"t3_n5f1s1","scoreUrl":"https://beatscratch.io/app/#/s/CZZX0"},{"filePath":null,"title":"A cheesy educational intro","author":"pseudocomposer","commentUrl":"https://reddit.com/r/BeatScratch/comments/myliqr/a_cheesy_educational_intro/","voteCount":1,"likes":true,"fullName":"t3_myliqr","scoreUrl":"https://beatscratch.io/app/#/s/ORXsf"},{"filePath":null,"title":"A longer, original demo using 5 instruments","author":"pseudocomposer","commentUrl":"https://reddit.com/r/BeatScratch/comments/my7ajg/a_longer_original_demo_using_5_instruments/","voteCount":1,"likes":true,"fullName":"t3_my7ajg","scoreUrl":"https://beatscratch.io/app/#/s/Z4hZh"},{"filePath":null,"title":"2021, From Jacob Collier’s Insta","author":"pseudocomposer","commentUrl":"https://reddit.com/r/BeatScratch/comments/mwyv7m/2021_from_jacob_colliers_insta/","voteCount":1,"likes":null,"fullName":"t3_mwyv7m","scoreUrl":"https://beatscratch.io/app/#/s/5dVNM"},{"filePath":null,"title":"Tee Time 2.6","author":"pseudocomposer","commentUrl":"https://reddit.com/r/BeatScratch/comments/lnmxyh/tee_time_26/","voteCount":1,"likes":null,"fullName":"t3_lnmxyh","scoreUrl":"https://beatscratch.io/app/#/s/rx0w0"}]';
@@ -70,7 +64,7 @@ class UniverseManager {
             ..putIfAbsent("likes", () => null))))
           .toList();
   List<ScoreFuture> get __cachedUniverseData =>
-      (_prefs?.getStringList('cachedUniverseData') ?? DEFAULT_UNIVERSE_DATA)
+      (_prefs.getStringList('cachedUniverseData') ?? DEFAULT_UNIVERSE_DATA)
           .map((it) => ScoreFuture.fromJson(jsonDecode(it)))
           .toList();
   List<ScoreFuture> _cachedUniverseData = [];
@@ -78,7 +72,7 @@ class UniverseManager {
   List<ScoreFuture> get cachedUniverseData => _cachedUniverseData;
   set cachedUniverseData(List<ScoreFuture> value) {
     _cachedUniverseData = value;
-    Future.microtask(() => _prefs?.setStringList("cachedUniverseData",
+    Future.microtask(() => _prefs.setStringList("cachedUniverseData",
         value.map((it) => jsonEncode(it.toJson())).toList()));
   }
 
@@ -91,8 +85,8 @@ class UniverseManager {
     refreshAccessToken(andPoll: true);
   }
 
-  String get _authState => _prefs?.getString('redditAuthState') ?? "";
-  set _authState(String value) => _prefs?.setString("redditAuthState", value);
+  String get _authState => _prefs.getString('redditAuthState') ?? "";
+  set _authState(String value) => _prefs.setString("redditAuthState", value);
 
   initiateSignIn() {
     _authState = uuid.v4();
@@ -110,9 +104,9 @@ class UniverseManager {
 
   bool tryAuthentication(String authUrl) {
     final uri = Uri.parse(authUrl);
-    String state = uri.queryParameters["state"];
-    String code = uri.queryParameters["code"];
-    if (state != null && code != null) {
+    String? state = uri.queryParameters["state"];
+    String? code = uri.queryParameters["code"];
+    if (code != null) {
       if (state != _authState) {
         messagesUI.sendMessage(
             message: "Auth codes did not match!",
@@ -139,7 +133,7 @@ class UniverseManager {
           final data = jsonDecode(response.body);
           String accessToken = data['access_token'];
           String refreshToken = data['refresh_token'];
-          if (accessToken != null && refreshToken != null) {
+          if (refreshToken != null) {
             redditRefreshToken = refreshToken;
             redditAccessToken = accessToken;
             loadRedditUsername();
@@ -170,10 +164,11 @@ class UniverseManager {
     redditAccessToken = "";
     redditUsername = "";
     redditRefreshToken = "";
-    cachedUniverseData =
-        cachedUniverseData.map((sf) => ScoreFuture.fromJson(sf.toJson()
+    cachedUniverseData = cachedUniverseData
+        .map((sf) => ScoreFuture.fromJson(sf.toJson()
           ..remove("likes")
-          ..putIfAbsent("likes", () => null)));
+          ..putIfAbsent("likes", () => null)))
+        .toList();
     refreshAccessToken();
   }
 
@@ -188,17 +183,13 @@ class UniverseManager {
       final username = data['name'];
       if (username != null) {
         redditUsername = username;
-        if (messagesUI != null) {
-          messagesUI.sendMessage(
-              message: "Reddit authentication successful!", andSetState: true);
-        }
+        messagesUI.sendMessage(
+            message: "Reddit authentication successful!", andSetState: true);
       } else {
-        if (messagesUI != null) {
-          messagesUI.sendMessage(
-              message: "Failed to load Reddit user information!",
-              isError: true,
-              andSetState: true);
-        }
+        messagesUI.sendMessage(
+            message: "Failed to load Reddit user information!",
+            isError: true,
+            andSetState: true);
       }
     });
   }
@@ -231,9 +222,7 @@ class UniverseManager {
     ).then((response) {
       final data = jsonDecode(response.body);
       String accessToken = data['access_token'];
-      if (accessToken != null) {
-        redditAccessToken = accessToken;
-      }
+      redditAccessToken = accessToken;
     });
   }
 
@@ -254,26 +243,29 @@ class UniverseManager {
     ).then((response) {
       final data = jsonDecode(response.body);
       String accessToken = data['access_token'];
-      if (accessToken != null) {
-        redditAccessToken = accessToken;
-      }
+      redditAccessToken = accessToken;
     });
   }
 
   Future<List<ScoreFuture>> loadUniverseData() async {
-    http.Response response = await http
-        .get(Uri.parse('https://oauth.reddit.com/r/BeatScratch/hot'),
-            headers: authenticatedRedditRequestHeaders)
-        .onError((error, stackTrace) {
-      print(error);
+    http.Response response = await http.get(
+        Uri.parse('https://oauth.reddit.com/r/BeatScratch/hot'),
+        headers: authenticatedRedditRequestHeaders);
+    //     .onError((error, stackTrace) {
+    //   print(error);
+    //   messagesUI.sendMessage(
+    //       message: "Error loading data from Reddit!",
+    //       isError: true,
+    //       andSetState: true);
+    //   // return null;
+    //   return Response(body, statusCode)
+    // });
+    if (response.statusCode != 200 && response.statusCode != 401) {
       messagesUI.sendMessage(
           message: "Error loading data from Reddit!",
           isError: true,
           andSetState: true);
-      return null;
-    });
-    if (response == null) {
-      return [];
+      return Future.value([]);
     }
     if (response.statusCode == 401) {
       await refreshAccessToken();
@@ -328,7 +320,7 @@ class UniverseManager {
     }
   }
 
-  vote(String fullName, bool likes, {bool andReauth = true}) async {
+  vote(String fullName, bool? likes, {bool andReauth = true}) async {
     http
         .post(Uri.parse('https://oauth.reddit.com/api/vote'),
             body: {
@@ -386,7 +378,7 @@ class UniverseManager {
         icon: Icon(FontAwesomeIcons.atom, color: chromaticSteps[0]),
         message: "Generating short URL via https://paste.ee...",
         andSetState: true);
-    String scoreUrl = await score.convertToShortUrl();
+    String? scoreUrl = await score.convertToShortUrl();
     messagesUI.sendMessage(
         icon: Icon(FontAwesomeIcons.atom, color: chromaticSteps[0]),
         message: "Uploading to the Universe...",
@@ -418,16 +410,12 @@ class UniverseManager {
         tryToSelectScore(int retries) {
           Future.delayed(Duration(seconds: 2), () {
             refreshUniverseData();
-            ScoreFuture scoreFuture = cachedUniverseData.firstWhere(
-                (it) => it.scoreUrl == scoreUrl,
-                orElse: () => null);
-            if (scoreFuture != null) {
-              messagesUI.setAppState(() {
-                currentUniverseScore = scoreFuture.identity;
-              });
-            } else if (retries > 0) {
-              tryToSelectScore(retries - 1);
-            }
+            final scoreFuture = cachedUniverseData
+                .firstWhereOrNull((it) => it.scoreUrl == scoreUrl);
+            if (scoreFuture == null) return;
+            messagesUI.setAppState(() {
+              currentUniverseScore = scoreFuture.identity;
+            });
           });
         }
 
