@@ -1,34 +1,13 @@
-import 'dart:async';
-import 'dart:ui';
-
 import '../music_preview/score_preview.dart';
-import 'package:dart_midi/dart_midi.dart';
-import 'package:dart_midi/src/byte_writer.dart';
 
-import '../messages/messages_ui.dart';
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/rendering.dart';
-import 'package:flutter_midi_command/flutter_midi_command.dart';
-import 'package:font_awesome_flutter/font_awesome_flutter.dart';
-import 'package:animated_list_plus/animated_list_plus.dart';
-import 'package:animated_list_plus/transitions.dart';
 
-import '../settings/settings.dart';
-import '../beatscratch_plugin.dart';
 import '../colors.dart';
-import '../generated/protos/protobeats_plugin.pb.dart';
 import '../generated/protos/protos.dart';
-import '../music_preview/melody_preview.dart';
 import '../storage/universe_manager.dart';
 import '../ui_models.dart';
-import '../universe_view/universe_icon.dart';
-import '../util/bs_methods.dart';
-import '../util/dummydata.dart';
-import '../util/midi_theory.dart';
 import '../util/util.dart';
 import '../widget/my_buttons.dart';
-import '../widget/my_platform.dart';
 
 showUniverseUpload(BuildContext context, Score score, Color sectionColor,
     UniverseManager universeManager, BSMethod onDoDuplicate) {
@@ -103,7 +82,10 @@ class UniverseUploadWidget extends StatefulWidget {
   final BSMethod onDoDuplicate;
 
   const UniverseUploadWidget(
-      {Key key, this.score, this.universeManager, this.onDoDuplicate})
+      {Key? key,
+      required this.score,
+      required this.universeManager,
+      required this.onDoDuplicate})
       : super(key: key);
 
   @override
@@ -111,7 +93,7 @@ class UniverseUploadWidget extends StatefulWidget {
 }
 
 class _UniverseUploadWidgetState extends State<UniverseUploadWidget> {
-  bool didFindDuplicate;
+  bool? didFindDuplicate;
   @override
   void initState() {
     super.initState();
