@@ -3,34 +3,26 @@ import 'package:beatscratch_flutter_redux/messages/messages_ui.dart';
 import 'package:beatscratch_flutter_redux/storage/universe_manager.dart';
 import 'package:beatscratch_flutter_redux/util/bs_methods.dart';
 import 'package:beatscratch_flutter_redux/widget/my_buttons.dart';
-
-import '../widget/my_popup_menu.dart';
 import 'package:flutter/material.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
-import 'dart:async';
-import 'dart:convert';
-import 'dart:io';
-import 'package:flutter/material.dart';
-import 'package:webview_flutter/webview_flutter.dart';
-import 'package:http/http.dart' as http;
-import 'dart:convert';
 
 import '../colors.dart';
 import '../ui_models.dart';
 import '../util/util.dart';
-import 'universe_icon.dart';
 import '../widget/my_platform.dart';
+import '../widget/my_popup_menu.dart';
+import 'universe_icon.dart';
 
 class UniverseViewUI {
-  BSMethod refreshUniverseData;
-  VoidCallback switchToLocalScores;
-  MessagesUI messagesUI;
+  BSMethod? refreshUniverseData;
+  VoidCallback? switchToLocalScores;
+  MessagesUI? messagesUI;
   bool visible = true;
   final UniverseManager universeManager;
   final Function(VoidCallback) setAppState;
   bool signingIn = false;
-  final Completer<WebViewController> _webViewController =
-      Completer<WebViewController>();
+  // final Completer<WebViewController> _webViewController =
+  //     Completer<WebViewController>();
   final TextEditingController _usernameController = TextEditingController();
   final TextEditingController _passwordController = TextEditingController();
 
@@ -41,25 +33,25 @@ class UniverseViewUI {
   }
 
   double toolbarHeight(BuildContext context,
-          {@required double keyboardHeight, @required double settingsHeight}) =>
+          {required double keyboardHeight, required double settingsHeight}) =>
       visible ? 44 : 0;
   double authFormHeight(BuildContext context,
-          {@required double keyboardHeight, @required double settingsHeight}) =>
+          {required double keyboardHeight, required double settingsHeight}) =>
       visible && signingIn ? 200 : 0;
   double height(BuildContext context,
-          {@required double keyboardHeight, @required double settingsHeight}) =>
+          {required double keyboardHeight, required double settingsHeight}) =>
       toolbarHeight(context,
           keyboardHeight: keyboardHeight, settingsHeight: settingsHeight) +
       authFormHeight(context,
           keyboardHeight: keyboardHeight, settingsHeight: settingsHeight);
 
   Widget build(
-      {@required BuildContext context,
-      @required Color sectionColor,
-      @required double keyboardHeight,
-      @required double settingsHeight,
-      @required VoidCallback showDownloads,
-      @required double scorePickerWidth}) {
+      {required BuildContext context,
+      required Color sectionColor,
+      required double keyboardHeight,
+      required double settingsHeight,
+      VoidCallback? showDownloads,
+      required double scorePickerWidth}) {
     double abbreviateAtWidth = 340;
     return AnimatedOpacity(
         duration: animationDuration,
@@ -90,7 +82,7 @@ class UniverseViewUI {
                 Transform.translate(
                   offset: Offset(0, 0),
                   child: MyFlatButton(
-                    onPressed: refreshUniverseData,
+                    onPressed: () => refreshUniverseData?.call(),
                     padding: EdgeInsets.all(5),
                     lightHighlight: true,
                     child: Transform.translate(
@@ -191,7 +183,7 @@ class UniverseViewUI {
                         case "signOut":
                           setAppState(() {
                             universeManager.signOut();
-                            messagesUI.sendMessage(
+                            messagesUI?.sendMessage(
                                 message: "Signed out of Reddit");
                           });
                           break;
