@@ -1,38 +1,33 @@
 import 'package:flutter/material.dart';
-import 'package:flutter/rendering.dart';
-import 'package:flutter/services.dart';
-import '../colors.dart';
 
 class MyFlatButton extends TextButton {
   MyFlatButton({
-    Key key,
-    @required VoidCallback onPressed,
-    VoidCallback onLongPress,
-    ValueChanged<bool> onHighlightChanged,
+    Key? key,
+    required VoidCallback? onPressed,
+    VoidCallback? onLongPress,
+    ValueChanged<bool>? onHighlightChanged,
     MouseCursor mouseCursor = SystemMouseCursors.basic,
-    ButtonTextTheme textTheme,
-    Color color,
-    EdgeInsetsGeometry padding,
+    ButtonTextTheme? textTheme,
+    Color? color,
+    EdgeInsetsGeometry? padding,
     Clip clipBehavior = Clip.none,
-    FocusNode focusNode,
+    FocusNode? focusNode,
     bool autofocus = false,
-    ButtonStyle style,
+    ButtonStyle? style,
     bool lightHighlight = false,
-    @required Widget child,
-  })  : assert(clipBehavior != null),
-        assert(autofocus != null),
-        super(
+    required Widget child,
+  }) : super(
           key: key,
           onPressed: onPressed,
           onLongPress: onLongPress,
           style: style ??
               ButtonStyle(
-                  backgroundColor: MaterialStateProperty.all(color),
-                  overlayColor: MaterialStateProperty.all(
+                  backgroundColor: WidgetStateProperty.all(color),
+                  overlayColor: WidgetStateProperty.all(
                       lightHighlight ? Colors.white10 : null),
                   mouseCursor:
-                      MaterialStateProperty.all(SystemMouseCursors.basic),
-                  padding: MaterialStateProperty.all(padding)),
+                      WidgetStateProperty.all(SystemMouseCursors.basic),
+                  padding: WidgetStateProperty.all(padding)),
           // ElevatedButton.styleFrom(
           //     primary: color,
           //     // onPrimary: color?.textColor(),
@@ -48,48 +43,46 @@ class MyFlatButton extends TextButton {
 
 class MyRaisedButton extends ElevatedButton {
   MyRaisedButton({
-    Key key,
-    @required VoidCallback onPressed,
-    VoidCallback onLongPress,
-    ValueChanged<bool> onHighlightChanged,
+    Key? key,
+    required VoidCallback? onPressed,
+    VoidCallback? onLongPress,
+    ValueChanged<bool>? onHighlightChanged,
     MouseCursor mouseCursor = SystemMouseCursors.basic,
-    ButtonTextTheme textTheme,
-    Color textColor,
-    Color disabledTextColor,
-    Color color,
-    Color disabledColor,
-    Color focusColor,
-    Color hoverColor,
-    Color highlightColor,
-    Color splashColor,
-    Brightness colorBrightness,
-    double elevation,
-    double focusElevation,
-    double hoverElevation,
-    double highlightElevation,
-    double disabledElevation,
-    EdgeInsetsGeometry padding,
-    VisualDensity visualDensity,
-    ShapeBorder shape,
+    ButtonTextTheme? textTheme,
+    Color? textColor,
+    Color? disabledTextColor,
+    Color? color,
+    Color? disabledColor,
+    Color? focusColor,
+    Color? hoverColor,
+    Color? highlightColor,
+    Color? splashColor,
+    Brightness? colorBrightness,
+    double elevation = 1.0,
+    double focusElevation = 1.0,
+    double hoverElevation = 1.0,
+    double highlightElevation = 1.0,
+    double disabledElevation = 1.0,
+    EdgeInsetsGeometry? padding,
+    VisualDensity? visualDensity,
+    ShapeBorder? shape,
     Clip clipBehavior = Clip.none,
-    FocusNode focusNode,
+    FocusNode? focusNode,
     bool autofocus = false,
-    MaterialTapTargetSize materialTapTargetSize,
-    Duration animationDuration,
-    Widget child,
-  })  : assert(autofocus != null),
-        assert(elevation == null || elevation >= 0.0),
-        assert(focusElevation == null || focusElevation >= 0.0),
-        assert(hoverElevation == null || hoverElevation >= 0.0),
-        assert(highlightElevation == null || highlightElevation >= 0.0),
-        assert(disabledElevation == null || disabledElevation >= 0.0),
-        assert(clipBehavior != null),
+    MaterialTapTargetSize? materialTapTargetSize,
+    Duration? animationDuration,
+    Widget? child,
+  })  : assert(elevation >= 0.0),
+        assert(focusElevation >= 0.0),
+        assert(hoverElevation >= 0.0),
+        assert(highlightElevation >= 0.0),
+        assert(disabledElevation >= 0.0),
         super(
           key: key,
           onPressed: onPressed,
           onLongPress: onLongPress,
           style: ElevatedButton.styleFrom(
-              primary: color,
+              backgroundColor: color,
               padding: padding,
               enabledMouseCursor: SystemMouseCursors.basic,
               disabledMouseCursor: SystemMouseCursors.basic),
@@ -104,20 +97,20 @@ class MyRaisedButton extends ElevatedButton {
 
 class MySlider extends Slider {
   const MySlider({
-    Key key,
-    @required double value,
-    @required ValueChanged<double> onChanged,
-    ValueChanged<double> onChangeStart,
-    ValueChanged<double> onChangeEnd,
+    Key? key,
+    required double value,
+    ValueChanged<double>? onChanged,
+    ValueChanged<double>? onChangeStart,
+    ValueChanged<double>? onChangeEnd,
     double min = 0.0,
     double max = 1.0,
-    int divisions,
-    String label,
-    Color activeColor,
-    Color inactiveColor,
+    int? divisions,
+    String? label,
+    Color? activeColor,
+    Color? inactiveColor,
     MouseCursor mouseCursor = SystemMouseCursors.basic,
-    SemanticFormatterCallback semanticFormatterCallback,
-    FocusNode focusNode,
+    SemanticFormatterCallback? semanticFormatterCallback,
+    FocusNode? focusNode,
     bool autofocus = false,
   }) : super(
           key: key,
