@@ -1,23 +1,20 @@
 import '../colors.dart';
 import '../music_view/music_action_button.dart';
-import 'package:flutter/cupertino.dart';
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/rendering.dart';
 
 import 'incrementable_value.dart';
 import '../ui_models.dart';
 
 class ScalableView extends StatefulWidget {
-  final VoidCallback onMicroScaleDown;
-  final VoidCallback onMicroScaleUp;
-  final VoidCallback onScaleDown;
-  final VoidCallback onScaleUp;
-  final Widget child;
+  final VoidCallback? onMicroScaleDown;
+  final VoidCallback? onMicroScaleUp;
+  final VoidCallback? onScaleDown;
+  final VoidCallback? onScaleUp;
+  final Widget? child;
   final String zoomLevelDescription; // e.g. 79%, 1x, 2x. Your choice.
   final bool autoScroll;
-  final VoidCallback toggleAutoScroll;
-  final VoidCallback scrollToCurrent;
+  final VoidCallback? toggleAutoScroll;
+  final VoidCallback? scrollToCurrent;
   final bool visible;
   final Color primaryColor;
   final bool showViewOptions;
@@ -25,12 +22,12 @@ class ScalableView extends StatefulWidget {
   final bool shiftUpZoomControls;
 
   const ScalableView(
-      {Key key,
+      {Key? key,
       this.onScaleDown,
       this.onScaleUp,
       this.child,
-      this.zoomLevelDescription,
-      this.autoScroll,
+      required this.zoomLevelDescription,
+      this.autoScroll = true,
       this.toggleAutoScroll,
       this.scrollToCurrent,
       this.visible = true,
@@ -47,128 +44,132 @@ class ScalableView extends StatefulWidget {
 }
 
 class _ScalableViewState extends State<ScalableView> {
-  double lastUpdatedScale;
-  int directionUpOrDown;
+  double? lastUpdatedScale;
+  int? directionUpOrDown;
   @override
   Widget build(BuildContext context) {
     return GestureDetector(
       child: Stack(
-        children: [
-          widget.child,
-          Row(children: [
-            Expanded(child: SizedBox()),
-            SizedBox(width: 2),
-            Column(children: [
-              Expanded(child: SizedBox()),
-              if (widget.autoScroll != null && widget.toggleAutoScroll != null)
-                MusicActionButton(
-                  visible: widget.visible && widget.showViewOptions,
-                  onPressed: widget.toggleAutoScroll,
-                  child: Stack(children: [
-                    Transform.translate(
-                        offset: Offset(0, -6),
-                        child: Text("Auto",
-                            maxLines: 1,
-                            overflow: TextOverflow.fade,
-                            style: TextStyle(
-                                fontSize: 10,
-                                color: widget.autoScroll
-                                    ? widget.primaryColor
-                                    : Colors.grey))),
-                    Transform.translate(
-                      offset: Offset(0, 6),
-                      child: AnimatedOpacity(
-                        duration: animationDuration,
-                        opacity: !widget.autoScroll ? 1 : 0,
-                        child:
-                            Icon(Icons.location_disabled, color: Colors.grey),
-                      ),
-                    ),
-                    Transform.translate(
-                      offset: Offset(0, 6),
-                      child: AnimatedOpacity(
-                        duration: animationDuration,
-                        opacity: widget.autoScroll ? 1 : 0,
-                        child:
-                            Icon(Icons.my_location, color: widget.primaryColor),
-                      ),
-                    ),
-                  ]),
-                ),
-              SizedBox(height: 2),
-              Container(
-                  color: Colors.black12,
-                  padding: EdgeInsets.all(0),
-                  child: IncrementableValue(
-                      child: Container(
-                        width: 48,
-                        height: 48,
-                        child: Align(
-                          alignment: Alignment.center,
-                          child: Transform.translate(
-                              offset: Offset(0, -3),
-                              child: Stack(children: [
-                                Transform.translate(
-                                    offset: Offset(-5, 5),
-                                    child: Transform.scale(
-                                        scale: 1,
-                                        child: Icon(Icons.zoom_out,
-                                            color: musicForegroundColor
-                                                .withOpacity(0.54)))),
-                                Transform.translate(
-                                    offset: Offset(5, -5),
-                                    child: Transform.scale(
-                                        scale: 1,
-                                        child: Icon(Icons.zoom_in,
-                                            color: musicForegroundColor
-                                                .withOpacity(0.54)))),
-                                Transform.translate(
-                                  offset: Offset(2, 20),
-                                  child: Text(widget.zoomLevelDescription,
-                                      style: TextStyle(
-                                          fontWeight: FontWeight.w800,
-                                          fontSize: 12,
-                                          color: musicForegroundColor
-                                              .withOpacity(0.87))),
-                                ),
-                              ])),
+        children:
+            ((widget.child != null ? [widget.child!] : []) as List<Widget>) +
+                [
+                  // widget.child,
+                  Row(children: [
+                    Expanded(child: SizedBox()),
+                    SizedBox(width: 2),
+                    Column(children: [
+                      Expanded(child: SizedBox()),
+                      if (widget.toggleAutoScroll != null)
+                        MusicActionButton(
+                          visible: widget.visible && widget.showViewOptions,
+                          onPressed: widget.toggleAutoScroll,
+                          child: Stack(children: [
+                            Transform.translate(
+                                offset: Offset(0, -6),
+                                child: Text("Auto",
+                                    maxLines: 1,
+                                    overflow: TextOverflow.fade,
+                                    style: TextStyle(
+                                        fontSize: 10,
+                                        color: widget.autoScroll
+                                            ? widget.primaryColor
+                                            : Colors.grey))),
+                            Transform.translate(
+                              offset: Offset(0, 6),
+                              child: AnimatedOpacity(
+                                duration: animationDuration,
+                                opacity: !widget.autoScroll ? 1 : 0,
+                                child: Icon(Icons.location_disabled,
+                                    color: Colors.grey),
+                              ),
+                            ),
+                            Transform.translate(
+                              offset: Offset(0, 6),
+                              child: AnimatedOpacity(
+                                duration: animationDuration,
+                                opacity: widget.autoScroll ? 1 : 0,
+                                child: Icon(Icons.my_location,
+                                    color: widget.primaryColor),
+                              ),
+                            ),
+                          ]),
                         ),
-                      ),
-                      valueWidth: 48,
-                      collapsing: true,
-                      musicActionButtonStyle: true,
-                      musicActionButtonColor:
-                          widget.zoomButtonColor.withOpacity(0.26),
-                      incrementIcon: Icons.zoom_in,
-                      decrementIcon: Icons.zoom_out,
-                      onIncrement: widget.onScaleUp,
-                      onDecrement: widget.onScaleDown)),
-              SizedBox(height: 2),
-              AnimatedContainer(
-                  duration: animationDuration,
-                  height: widget.shiftUpZoomControls ? 24 : 0)
-            ]),
-            SizedBox(width: 2),
-          ])
-        ],
+                      SizedBox(height: 2),
+                      Container(
+                          color: Colors.black12,
+                          padding: EdgeInsets.all(0),
+                          child: IncrementableValue(
+                              child: Container(
+                                width: 48,
+                                height: 48,
+                                child: Align(
+                                  alignment: Alignment.center,
+                                  child: Transform.translate(
+                                      offset: Offset(0, -3),
+                                      child: Stack(children: [
+                                        Transform.translate(
+                                            offset: Offset(-5, 5),
+                                            child: Transform.scale(
+                                                scale: 1,
+                                                child: Icon(Icons.zoom_out,
+                                                    color: musicForegroundColor
+                                                        .withOpacity(0.54)))),
+                                        Transform.translate(
+                                            offset: Offset(5, -5),
+                                            child: Transform.scale(
+                                                scale: 1,
+                                                child: Icon(Icons.zoom_in,
+                                                    color: musicForegroundColor
+                                                        .withOpacity(0.54)))),
+                                        Transform.translate(
+                                          offset: Offset(2, 20),
+                                          child: Text(
+                                              widget.zoomLevelDescription,
+                                              style: TextStyle(
+                                                  fontWeight: FontWeight.w800,
+                                                  fontSize: 12,
+                                                  color: musicForegroundColor
+                                                      .withOpacity(0.87))),
+                                        ),
+                                      ])),
+                                ),
+                              ),
+                              valueWidth: 48,
+                              collapsing: true,
+                              musicActionButtonStyle: true,
+                              musicActionButtonColor:
+                                  widget.zoomButtonColor.withOpacity(0.26),
+                              incrementIcon: Icons.zoom_in,
+                              decrementIcon: Icons.zoom_out,
+                              onIncrement: widget.onScaleUp,
+                              onDecrement: widget.onScaleDown)),
+                      SizedBox(height: 2),
+                      AnimatedContainer(
+                          duration: animationDuration,
+                          height: widget.shiftUpZoomControls ? 24 : 0)
+                    ]),
+                    SizedBox(width: 2),
+                  ])
+                ],
       ),
       onScaleStart: (details) {
         lastUpdatedScale = 1;
       },
       onScaleUpdate: (details) {
+        final lastUpdatedScale = this.lastUpdatedScale;
+        if (lastUpdatedScale == null) {
+          return;
+        }
         final scale = details.scale;
         // print("scale: $scale");
-        if (widget.onMicroScaleUp != null && scale - lastUpdatedScale >= .01) {
-          widget.onMicroScaleUp();
-        } else if (widget.onMicroScaleDown != null &&
-            scale - lastUpdatedScale <= -.01) {
-          widget.onMicroScaleDown();
-        } else if (widget.onScaleUp != null &&
-            scale - lastUpdatedScale >= .09) {
-          widget.onScaleUp();
-        } else if (widget.onScaleDown != null &&
-            scale - lastUpdatedScale <= -.09) {
-          widget.onScaleDown();
+        if (scale - lastUpdatedScale >= .01) {
+          widget.onMicroScaleUp?.call();
+        } else if (scale - lastUpdatedScale <= -.01) {
+          widget.onMicroScaleDown?.call();
+        } else if (scale - lastUpdatedScale >= .09) {
+          widget.onScaleUp?.call();
+        } else if (scale - lastUpdatedScale <= -.09) {
+          widget.onScaleDown?.call();
         }
       },
       onScaleEnd: (details) {
