@@ -3,7 +3,6 @@ import 'dart:collection';
 import '../beatscratch_plugin.dart';
 import '../generated/protos/music.pb.dart';
 import '../generated/protos/protobeats_plugin.pb.dart';
-import '../util/bs_methods.dart';
 import '../util/util.dart';
 import '../util/proto_utils.dart';
 import '../util/music_theory.dart';
@@ -13,11 +12,11 @@ import '../util/music_utils.dart';
 /// are sent to this singleton queue for processing.
 class RecordedSegmentQueue {
   /// Gets the recording melody from your UI. Should be set in [State.initState] and [State.dispose].
-  static Melody Function() getRecordingMelody;
+  static Melody? Function()? getRecordingMelody;
 
   /// Should be set in [State.initState] and [State.dispose] for
-  static Function(Melody) updateRecordingMelody;
-  static Melody get recordingMelody => getRecordingMelody?.call();
+  static Function(Melody)? updateRecordingMelody;
+  static Melody? get recordingMelody => getRecordingMelody?.call();
   // static set recordingMelody(Melody melody) => updateRecordingMelody(melody);
   static final Queue<RecordedSegment> segments = ListQueue<RecordedSegment>();
   static final BSValueMethod<bool> enabled = BSValueMethod(false)
@@ -49,20 +48,20 @@ class RecordedSegmentQueue {
   static _processSegment(RecordedSegment segment) {
     if (segment.recordedData.isEmpty) return;
     final melody = recordingMelody;
-    if (melody != null) {
-      RecordedSegment_RecordedBeat firstBeat =
-          segment.beats.minBy((rb) => rb.timestamp.toInt());
-      RecordedSegment_RecordedBeat secondBeat =
-          segment.beats.maxBy((rb) => rb.timestamp.toInt());
-      segment.recordedData.forEach((data) {
-        _processSegmentData(segment, data, melody, firstBeat, secondBeat);
-      });
-      print("applying PostProcessing: separateNoteOnAndOffs()");
-      melody.separateNoteOnAndOffs();
-      print("updateRecordingMelody?.call: ${melody.logString}");
-      updateRecordingMelody?.call(melody);
-      BeatScratchPlugin.updateMelody(melody);
-    }
+    if (melody == null) return;
+
+    RecordedSegment_RecordedBeat firstBeat =
+        segment.beats.minBy((rb) => rb.timestamp.toInt());
+    RecordedSegment_RecordedBeat secondBeat =
+        segment.beats.maxBy((rb) => rb.timestamp.toInt())!;
+    segment.recordedData.forEach((data) {
+      _processSegmentData(segment, data, melody, firstBeat, secondBeat);
+    });
+    print("applying PostProcessing: separateNoteOnAndOffs()");
+    melody.separateNoteOnAndOffs();
+    print("updateRecordingMelody?.call: ${melody.logString}");
+    updateRecordingMelody?.call(melody);
+    BeatScratchPlugin.updateMelody(melody);
   }
 
   static _processSegmentData(
