@@ -1,16 +1,16 @@
 import 'package:protobuf/protobuf.dart';
 
-import 'fake_js.dart'
-if(dart.library.js) 'dart:js';
+import 'fake_js.dart' if (dart.library.js) 'dart:js';
 
 extension ProtoUtils<T extends GeneratedMessage> on T {
-  dynamic protoJsify() => JsObject.jsify(bsCopy().toProto3Json());
+  dynamic protoJsify() => JsObject.jsify(bsCopy().toProto3Json()!);
   T bsCopy() {
     return deepCopy();
   }
+
   T bsRebuild(Function(T) updates) {
-    return (deepCopy()..freeze()).rebuild((t) => updates(t)).deepCopy() as T;
+    return (deepCopy()..freeze()).rebuild((t) => updates(t)).deepCopy();
   }
 
-  String get logString =>  "\n  ${toString().replaceAll("\n", "\n  ")}";
+  String get logString => "\n  ${toString().replaceAll("\n", "\n  ")}";
 }
